% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-Data.R
\docType{data}
\name{Beretvas2008}
\alias{Beretvas2008}
\alias{Borckardt2014}
\alias{Huitema2000}
\alias{Waddell2011}
\alias{byHeart2011}
\alias{Grosche2011}
\alias{Grosche2014}
\alias{GruenkeWilbert2014}
\alias{Huber2014}
\alias{example_A24}
\alias{exampleAB_simple}
\alias{exampleAB}
\alias{exampleAB_add}
\alias{exampleAB_decreasing}
\alias{exampleABC}
\alias{exampleABC_outlier}
\alias{exampleABC_50}
\alias{exampleABAB}
\alias{exampleA1B1A2B2}
\alias{exampleA1B1A2B2_zvt}
\alias{exampleAB_50}
\alias{exampleAB_50.l2}
\alias{exampleABC_150}
\alias{Lenz2013}
\alias{Leidig2018}
\alias{Leidig2018_l2}
\alias{SSDforR2017}
\alias{Parker2011}
\title{Single-case example data}
\format{
An object of class \code{scdf} (inherits from \code{list}) of length 1.
}
\usage{
Beretvas2008
}
\value{
\item{Beretvas2008}{Fictitious single-case intervention study.
\bold{Reference:} Beretvas, S., & Chung, H. (2008). An evaluation of
modified R2-change effect size indices for single-subject experimental
designs. \emph{Evidence-Based Communication Assessment and Intervention, 2},
120-128.} 
\item{Borckardt2014}{Fictitious daily pain ratings evaluating a
psychological treatment. \bold{Reference:} Borckardt, J. J., & Nash, M. R.
(2014). Simulation modelling analysis for small sets of single-subject data
collected over time. \emph{Neuropsychological Rehabilitation, 24}, 492-506.}
\item{Huitema2000}{Fictitious single-case intervention study.
\bold{Reference:} Huitema, B. E., & McKean, J. W. (2000). Design
specification issues in time-series intervention models. \emph{Educational
and Psychological Measurement, 60}, 38-58.} 
\item{Waddell2011}{Fictitious
single-case intervention study. \bold{Reference:} Waddell, D. E., Nassar, S.
L., & Gustafson, S. A. (2011). Single-Case Design in Psychophysiological
Research: Part II: Statistical Analytic Approaches. \emph{Journal of
Neurotherapy, 15}, 160-169.} 
\item{byHeart2011}{Multiple-baseline (11 cases)
intervention study on flash card vocabulary learning by Juergen Wilbert.}
\item{Grosche2011}{Multiple-baseline (3 cases) intervention study on a
direct-instructive reading intervention. \bold{Reference:} Grosche, M.
(2011). Effekte einer direkt-instruktiven Foerderung der Lesegenauigkeit.
\emph{Empirische Sonderpaedagogik, 3}, 147-161.}
\item{Grosche2014}{Multiple-baseline (3 cases x 3 materials) intervention
study on a reading intervention. \bold{Reference:} Grosche, M., Lueke, T., &
Wilbert, J. (in prep.).} 
\item{GruenkeWilbert2014}{Multiple-baseline (6
cases) intervention study on story mapping. \bold{Reference:} Gruenke, M.,
Wilbert, J., & Stegemann-Calder, K. (2013). Analyzing the effects of story
mapping on the reading comprehension of children with low intellectual
abilities. \emph{Learning Disabilities: A Contemporary Journal, 11}, 51-64.}
\item{Huber2014}{Multiple-baseline (4 cases) intervention study on
behavioral compliance. Scores refer to compliant behavior in percent.
\bold{Reference:} Huber, C. (in prep.).}
\item{Lenz2013}{Fictious example from the paper Lenz, A. S. (2013). Calculating Effect Size in Single-Case Research: A Comparison of Nonoverlap Methods. Measurement and Evaluation in Counseling and Development, 46(1), 64–73.}
\item{Leidig2018}{}
\item{Leidig2018_l2}{}
\item{SSDforR2017}{Example from the R package SSDforR.}
\item{Parker2011}{Example from Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. (2011). Combining Nonoverlap and Trend for Single-Case Research: Tau-U. Behavior Therapy, 42(2), 284–299. https://doi.org/10.1016/j.beth.2010.08.006}
}
\description{
The \code{scan} package comes with a set of fictitious and authentic
single-case study data, by courtesy of the particular authors.
}
\author{
Juergen Wilbert
}
\keyword{datasets}
