% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf-functions.R
\name{c.scdf}
\alias{c.scdf}
\title{Concatenate single-case data frames}
\usage{
\method{c}{scdf}(...)
}
\arguments{
\item{...}{scdf objects}
}
\value{
A scdf
}
\description{
Concatenate single-case data frames
}
