% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scdf.R
\name{check_scdf}
\alias{check_scdf}
\title{Validity check for an scdf object}
\usage{
check_scdf(object)
}
\arguments{
\item{object}{An scdf object}
}
\value{
TRUE or list with error and warning messages.
}
\description{
Validity check for an scdf object
}
\examples{
check_scdf(exampleAB)

check_scdf(c(exampleAB, exampleABC))

}
