% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R, R/deprecated.R
\name{code}
\alias{code}
\alias{convert}
\title{Code
Converts an scdf object into R code}
\usage{
code(scdf, file = "", study_name = "study", case_name = "case", indent = 3)

convert(...)
}
\arguments{
\item{scdf}{A single-case data frame. See \code{\link{scdf}} to learn about
this format.}

\item{file}{A filename for exporting the syntax.}

\item{study_name}{Character string. Name of the study object.}

\item{case_name}{Character string. Name of the scdf objects.}

\item{indent}{Integer. Indentation.}

\item{...}{Further arguments passed to the function.}
}
\description{
Code
Converts an scdf object into R code
}
\examples{
filename <- tempfile()
code(exampleABC, file = filename)
source(filename)
all.equal(study, exampleABC)
unlink(filename)
}
