% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf-functions.R
\name{combine}
\alias{combine}
\alias{c.scdf}
\title{Combine single-case data frames into a multiple case study}
\usage{
combine(...)

\method{c}{scdf}(...)
}
\arguments{
\item{...}{scdf objects}
}
\value{
A scdf
}
\description{
Combine single-case data frames into a multiple case study
}
