% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_tauSC.R
\name{corrected_tauSC}
\alias{corrected_tauSC}
\title{Baseline corrected tau}
\usage{
corrected_tauSC(data, dvar, pvar, mvar, phases = c(1, 2), alpha = 0.05,
  continuity = TRUE, repeated = TRUE)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{dvar}{Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.}

\item{pvar}{Character string with the name of the phase variable. Defaults to the attributes in the scdf file.}

\item{mvar}{Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.}

\item{phases}{A vector of two characters or numbers indicating the two
phases that should be compared. E.g., \code{phases = c("A","C")} or
\code{phases = c(2,4)} for comparing the second to the fourth phase. Phases
could be combined by providing a list with two elements. E.g., \code{phases
= list(A = c(1,3), B = c(2,4))} will compare phases 1 and 3 (as A) against 2
and 4 (as B). Default is \code{phases = c("A","B")}.}

\item{alpha}{Sets the p-value at and below which a baseline correction is applied.}

\item{continuity}{If TRUE applies a continuity correction for calculating p}

\item{repeated}{If TRUE applies the repeated median method for caluclating slope and intercept (\code{\link{mblm}})}
}
\description{
Kendalls tau correlation for the dependent variable and the phase variable is calculated
after correcting for a baseline trend.
}
\details{
This method has been proposed by Tarlow (2016). The baseline data are checked for a singificant 
autocorrelation (based on Kendalls Tau). If so, a non-parameteric Theil-Sen regression is applied
for the baseline data where the dependent values are regressed on the measurement time. The resulting slope
information is then used to predict data of the B-phase. The dependent variable is now corrected for this baseline trend 
and the resudials of the Theil-Sen regression are taken for further caluculations.
Finally, a tau is calculated for the dependent variable and the dichtomos phase variable.
The function here provides two extensions to this procedure: The more accurate Siegel repeated median regression
is applied when \code{repeated = TRUE} and a continuity correction is applied when \code{continuity = TRUE} (both are the default settings).
}
\examples{
dat <- scdf(c(A = 33,25,17,25,14,13,15, B = 15,16,16,5,7,9,6,5,3,3,8,11,7))
corrected_tauSC(dat)
}
\references{
Tarlow, K. R. (2016). An Improved Rank Correlation Effect 
Size Statistic for Single-Case Designs: Baseline Corrected Tau. Behavior Modification, 41(4), 427–467. https://doi.org/10.1177/0145445516676750
}
\seealso{
Other regression functions: \code{\link{hplm}},
  \code{\link{mplm}}, \code{\link{plm}}

Other overlap functions: \code{\link{nap}},
  \code{\link{overlapSC}}, \code{\link{pand}},
  \code{\link{pem}}, \code{\link{pet}}, \code{\link{pnd}},
  \code{\link{tauUSC}}
}
\concept{overlap functions}
\concept{regression functions}
