% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.scdf.R, R/autocorr.R,
%   R/corrected_tau.R, R/deprecated.R, R/describe.R, R/design.R,
%   R/fill_missing.R, R/outlier.R, R/overlap.R, R/power_test.R, R/rand_test.R,
%   R/random_scdf.R, R/ranks.R, R/rci.R, R/shift.R, R/smooth_cases.R,
%   R/style_plot.R, R/tau_u.R, R/trend.R, R/truncate_phase.R
\name{longSCDF}
\alias{longSCDF}
\alias{autocorrSC}
\alias{corrected_tauSC}
\alias{deprecated-functions}
\alias{describeSC}
\alias{design_rSC}
\alias{fillmissingSC}
\alias{outlierSC}
\alias{overlapSC}
\alias{power_testSC}
\alias{randSC}
\alias{rand.test}
\alias{rSC}
\alias{rankSC}
\alias{rciSC}
\alias{rCi}
\alias{shiftSC}
\alias{smoothSC}
\alias{style_plotSC}
\alias{style.plotSC}
\alias{tauUSC}
\alias{trendSC}
\alias{truncateSC}
\title{List of old deprecated function names}
\usage{
longSCDF(...)

autocorrSC(...)

corrected_tauSC(...)

describeSC(...)

design_rSC(...)

fillmissingSC(...)

outlierSC(...)

overlapSC(...)

power_testSC(...)

randSC(...)

rand.test(...)

rSC(...)

rankSC(...)

rciSC(...)

rCi(...)

shiftSC(...)

smoothSC(...)

style_plotSC(...)

style.plotSC(...)

tauUSC(...)

trendSC(...)

truncateSC(...)
}
\arguments{
\item{...}{Arguments passed to the design function.}
}
\description{
This is a list with functions names that have been replaced by new names
which are in line with re R syntax guidelines. The old function names still 
work. They are wrappers that call the new function.
}
