% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R, R/print-export-desc.R
\name{describe}
\alias{describe}
\alias{print.sc_desc}
\alias{export.sc_desc}
\title{Descriptive statistics for single-case data}
\usage{
describe(data, dvar, pvar, mvar)

\method{print}{sc_desc}(x, digits = "auto", ...)

\method{export}{sc_desc}(
  object,
  caption = NA,
  footnote = NA,
  filename = NA,
  flip = FALSE,
  round = 2,
  ...
)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link[=scdf]{scdf()}} to learn about this
format.}

\item{dvar}{Character string with the name of the dependent variable.
Defaults to the attributes in the scdf file.}

\item{pvar}{Character string with the name of the phase variable. Defaults to
the attributes in the scdf file.}

\item{mvar}{Character string with the name of the measurement time variable.
Defaults to the attributes in the scdf file.}

\item{x}{Object}

\item{digits}{The minimum number of significant digits to be use.
If set to "auto" (default), values are predefined.}

\item{...}{Further arguments passed to the function.}

\item{object}{An scdf or an object exported from a scan function.}

\item{caption}{Character string with table caption. If left NA (default) a
caption will be created based on the exported object.}

\item{footnote}{Character string with table footnote. Several strings are
combined into a footnote of several lines. If left NA (default) a footnote
will be created based on the exported object. \code{NULL} or \code{""} suppress the
footnote.}

\item{filename}{String containing the file name. If a filename is given the
output will be written to that file.}

\item{flip}{If TRUE, some objects are exported with rows and columns flipped.}

\item{round}{Integer passed to the digits argument used to round values.}
}
\value{
A list containing a data frame of descriptive statistics
(descriptives); the cse design (design); the number of cases (N).
}
\description{
The \code{describe()} function provides common descriptive statistics for
single-case data.
}
\details{
It computes the number of measurements, number of missing values, mean,
median, standard deviation, median average deviation, minimum, maximum, and
trend (slope of dependent variable regressed on measurement-time) for each
phase of each single-case included in an scdf.

n = number of measurements; mis = number of missing vaues; m = mean;
md = median; sd = standard deviation; mad = median average deviation; min =
minimum; max = maximum; trend = weight of depended variable regressed on
time (values ~ mt).
}
\section{Functions}{
\itemize{
\item \code{print(sc_desc)}: Print results

\item \code{export(sc_desc)}: Export results as html table (see \code{\link[=export]{export()}})

}}
\examples{

## Descriptive statistics for a study of three single-cases
describe(Grosche2011)

## Descriptives of a three phase design
describe(exampleABC)

## Write descriptive statistics to .csv-file
study <- describe(Waddell2011)
write.csv(study$descriptives, file = tempfile())
}
\seealso{
\code{\link[=overlap]{overlap()}}, \code{\link[=plot.scdf]{plot.scdf()}}
}
\author{
Juergen Wilbert
}
