% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit_roxygen_descriptions.R
\name{.inheritParams}
\alias{.inheritParams}
\title{Dummy function to inherit global descriptions of parameters}
\usage{
.inheritParams(
  data,
  scdf,
  dvar,
  mvar,
  pvar,
  decreasing,
  phases,
  model,
  contrast,
  contrast_level,
  contrast_slope,
  trend,
  level,
  slope,
  nice,
  flip,
  decimals,
  select,
  ...
)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link[=scdf]{scdf()}} to learn about this
format.}

\item{scdf}{A single-case data frame. See \code{\link[=scdf]{scdf()}} to learn about this
format.}

\item{dvar}{Character string with the name of the dependent variable.
Defaults to the attributes in the scdf file.}

\item{mvar}{Character string with the name of the measurement time variable.
Defaults to the attributes in the scdf file.}

\item{pvar}{Character string with the name of the phase variable. Defaults to
the attributes in the scdf file.}

\item{decreasing}{If you expect data to be lower in the B phase, set
\code{decreasing = TRUE}. Default is \code{decreasing = FALSE}.}

\item{phases}{A vector of two characters or numbers indicating the two phases
that should be compared. E.g., \code{phases = c("A","C")} or \code{phases = c(2,4)}
for comparing the second to the fourth phase. Phases could be combined by
providing a list with two elements. E.g., \code{phases = list(A = c(1,3), B = c(2,4))} will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default
is \code{phases = c(1,2)}.}

\item{model}{Model used for calculating the dummy parameters (see Huitema &
McKean, 2000). Default is \code{model = "W"}. Possible values are: \code{"B&L-B"},
\code{"H-M"}, \code{"W"}, and deprecated \code{"JW"}.}

\item{contrast}{Sets contrast_level and contrast_slope. Either "first",
"preceding" or a contrast matrix. If NA contrast is ignored.}

\item{contrast_level}{Either "first", "preceding" or a contrast matrix. If NA
contrast_level is a copy of contrast.}

\item{contrast_slope}{Either "first", "preceding" or a contrast matrix. If NA
contrast_level is a copy of contrast.}

\item{trend}{A logical indicating if a trend parameters is included in the
model.}

\item{level}{A logical indicating if a level parameters is included in the
model.}

\item{slope}{A logical indicating if a slope parameters is included in the
model.}

\item{nice}{If set TRUE (default) output values are rounded and optimized for
publication tables.}

\item{flip}{If TRUE, some objects are exported with rows and columns flipped.}

\item{decimals}{Decimal places that are reported.}

\item{select}{A character vector containing the names of the variables to be
included. If the vector is named, the variables will be renamed
accordingly.}

\item{...}{Further arguments passed to the function.}
}
\value{
No return value.
}
\description{
This function is only used to inherit parameter descriptions in the
documentation of other functions. It has no other purpose and is not meant to
be called directly.
}
\author{
Juergen Wilbert
}
\concept{internal functions}
\keyword{internal}
