% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw random names}
\usage{
draw_names(n = 1, type = "neutral", seed = NULL)
}
\arguments{
\item{n}{Number of names}

\item{type}{"neutral", "male", "female", or "mixed"}

\item{seed}{A seed for the random number generator.}
}
\value{
A character vector with random names
}
\description{
Draw random names
}
