\name{example-data}
\alias{example}
\alias{Beretvas2008}
\alias{Borckardt2014}
\alias{Huitema2000}
\alias{Waddell2011}
\alias{byHeart2011}
\alias{example1}
\alias{Grosche2011}
\alias{Grosche2014}
\alias{GruenkeWilbert2014}
\alias{Huber2014}
\title{Single-case example data}
\description{
The \code{scan} package comes with a set of fictitious and authentic single-case study data, by courtesy of the particular authors.
}

\value{
  \item{Beretvas2008}{Fictitious single-case intervention study. \bold{Reference:} Beretvas, S., & Chung, H. (2008). An evaluation of modified R2-change effect size indices for single-subject experimental designs. \emph{Evidence-Based Communication Assessment and Intervention, 2}, 120-128.}
  \item{Borckardt2014}{Fictitious daily pain ratings evaluating a psychological treatment. \bold{Reference:} Borckardt, J. J., & Nash, M. R. (2014). Simulation modelling analysis for small sets of single-subject data collected over time. \emph{Neuropsychological Rehabilitation, 24}, 492-506.}
  \item{Huitema2000}{Fictitious single-case intervention study. \bold{Reference:} Huitema, B. E., & McKean, J. W. (2000). Design specification issues in time-series intervention models. \emph{Educational and Psychological Measurement, 60}, 38-58.}
  \item{Waddell2011}{Fictitious single-case intervention study. \bold{Reference:} Waddell, D. E., Nassar, S. L., & Gustafson, S. A. (2011). Single-Case Design in Psychophysiological Research: Part II: Statistical Analytic Approaches. \emph{Journal of Neurotherapy, 15}, 160-169.}
  \item{byHeart2011}{Multiple-baseline (11 cases) intervention study on flash card vocabulary learning by Juergen Wilbert.}
  \item{example1}{Fictitious multiple-baseline (3 cases) design.}
  \item{Grosche2011}{Multiple-baseline (3 cases) intervention study on a direct-instructive reading intervention. \bold{Reference:} Grosche, M. (2011). Effekte einer direkt-instruktiven Foerderung der Lesegenauigkeit. \emph{Empirische Sonderpaedagogik, 3}, 147-161.}
  \item{Grosche2014}{Multiple-baseline (3 cases x 3 materials) intervention study on a reading intervention. \bold{Reference:} Grosche, M., Lueke, T., & Wilbert, J. (in prep.).}
  \item{GruenkeWilbert2014}{Multiple-baseline (6 cases) intervention study on story mapping. \bold{Reference:} Gruenke, M., Wilbert, J., & Stegemann-Calder, K. (2013). Analyzing the effects of story mapping on the reading comprehension of children with low intellectual abilities. \emph{Learning Disabilities: A Contemporary Journal, 11}, 51-64.}
  \item{Huber2014}{Multiple-baseline (4 cases) intervention study on behavioral compliance. Scores refer to compliant behavior in percent. \bold{Reference:} Huber, C. (in prep.).}
}

\author{Juergen Wilbert}