% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R, R/print-export-scdf.R
\name{export}
\alias{export}
\alias{export.scdf}
\title{Export scan objects to html or latex}
\usage{
export(object, ...)

\method{export}{scdf}(
  object,
  summary = FALSE,
  caption = NA,
  footnote = NA,
  filename = NA,
  cols,
  round = 2,
  ...
)
}
\arguments{
\item{object}{An scdf or an object exported from a scan function.}

\item{...}{Further Arguments passed to internal functions.}

\item{summary}{If TRUE, exports the summary of an \code{scdf}.}

\item{caption}{Character string with table caption. If left NA (default) a
caption will be created based on the exported object.}

\item{footnote}{Character string with table footnote. Several strings are
combined into a footnote of several lines. If left NA (default) a footnote
will be created based on the exported object. \code{NULL} or \code{""} suppress the
footnote.}

\item{filename}{String containing the file name. If a filename is given the
output will be written to that file.}

\item{cols}{Defines which columns are included when exporting an scdf. It is
either a vector of variable names or the string "main" will select the
central variables.}

\item{round}{Integer passed to the digits argument used to round values.}
}
\value{
Returns or displays a specially formatted html (or latex) file.
}
\description{
Export creates html files of tables or displays them directly in the viewer
pane of rstudio. When applied in rmarkdown/quarto, tables can also be created
for pdf/latex output.
}
\details{
The function uses either the gt or the kableExtra package to create the
tables. The default engine is gt; in a document knitted to latex, kableExtra
is used instead, because gt tables receive neither a number nor a label
there, and for Word output gt is used even when kableExtra is set, because
kableExtra can not write Word tables. The engine can be set via the option
\code{scan.export.engine}. Additional
options for kable and kable_styling can be set via the options
\code{scan.export.kable} and \code{scan.export.kable_styling}.
The default options can be viewed and modified via
\code{options("scan.export.kable")} and
\code{options("scan.export.kable_styling")}.
}
