% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_scplot.R
\name{install_scplot}
\alias{install_scplot}
\title{Installation instructions for scplot}
\usage{
install_scplot()
}
\description{
scplot is a new extended plotting interface for scan based on ggplot2.
It is not yet implemented into scan or available on cran and must be 
installed from github. This functions gives instructions how to do this.
}
