% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_shinyscan.R
\name{install_shinyscan}
\alias{install_shinyscan}
\title{Installation instructions for shiny}
\usage{
install_shinyscan()
}
\description{
shinyscan is a graphical user interface for scan based on shiny.
It is not yet implemented into scan or available on cran and must be 
installed from github along with the scplot package.
}
