% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf-functions.R
\name{is.scdf}
\alias{is.scdf}
\title{Test for scdf objects}
\usage{
is.scdf(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Returns TRUE or FALSE depending on whether its argument is of scdf type or not.
}
\description{
Tests for objects of type "scdf".
}
\author{
Juergen Wilbert
}
