% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_function.R
\name{mc_function}
\alias{mc_function}
\title{Returns a function to use in Power analysis and Monte-Carlo studies}
\usage{
mc_function(name)
}
\arguments{
\item{name}{Character string. Identifier of function.}
}
\value{
A function.
}
\description{
Returns a function to use in Power analysis and Monte-Carlo studies
}
\details{
If 'name' is unknown, a list with valid names is provided.
}
