% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet.R, R/print-export-pet.R
\name{pet}
\alias{pet}
\alias{print.sc_pet}
\alias{export.sc_pet}
\title{Percent exceeding the trend (PET)}
\usage{
pet(data, dvar, pvar, mvar, ci = 0.95, decreasing = FALSE, phases = c(1, 2))

\method{print}{sc_pet}(x, digits = 3, ...)

\method{export}{sc_pet}(object, caption = NA, footnote = NA, filename = NA, round = 1, ...)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link[=scdf]{scdf()}} to learn about this
format.}

\item{dvar}{Character string with the name of the dependent variable.
Defaults to the attributes in the scdf file.}

\item{pvar}{Character string with the name of the phase variable. Defaults to
the attributes in the scdf file.}

\item{mvar}{Character string with the name of the measurement time variable.
Defaults to the attributes in the scdf file.}

\item{ci}{Width of the confidence interval. Default is \code{ci = 0.95}.}

\item{decreasing}{If you expect data to be lower in the B phase, set
\code{decreasing = TRUE}. Default is \code{decreasing = FALSE}.}

\item{phases}{A vector of two characters or numbers indicating the two phases
that should be compared. E.g., \code{phases = c("A","C")} or \code{phases = c(2,4)}
for comparing the second to the fourth phase. Phases could be combined by
providing a list with two elements. E.g., \code{phases = list(A = c(1,3), B = c(2,4))} will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default
is \code{phases = c(1,2)}.}

\item{x}{Object}

\item{digits}{The minimum number of significant digits to be use.
If set to "auto" (default), values are predefined.}

\item{...}{Further arguments passed to the function.}

\item{object}{An scdf or an object exported from a scan function.}

\item{caption}{Character string with table caption. If left NA (default) a
caption will be created based on the exported object.}

\item{footnote}{Character string with table footnote. Several strings are
combined into a footnote of several lines. If left NA (default) a footnote
will be created based on the exported object. \code{NULL} or \code{""} suppress the
footnote.}

\item{filename}{String containing the file name. If a filename is given the
output will be written to that file.}

\item{round}{Integer passed to the digits argument used to round values.}
}
\value{
\tabular{ll}{
    \tab  \cr
   \code{PET} \tab Percent exceeding the trend. \cr
   \code{ci} \tab Width of confidence interval. \cr
   \code{decreasing} \tab Logical argument from function call (see Arguments above). \cr
}
}
\description{
The \code{pet} function returns the percentage of Phase B data points that exceed
the prediction based on the Phase A trend. A binomial test against a 50/50
distribution is calculated. It also calculates the percentage of Phase B data
points that exceed the upper (or lower) 95 percent confidence interval of the
predicted progression.
}
\details{
The PET is calculated by first fitting a linear model to the Phase A data
to estimate the trend. Then, for each data point in Phase B, it is
determined whether it exceeds the predicted value from the Phase A trend.
The PET is the percentage of Phase B data points that exceed this predicted
value. Additionally, a binomial test is performed to assess whether the
observed PET is significantly greater than what would be expected by chance
(i.e., 50\%).
}
\section{Functions}{
\itemize{
\item \code{print(sc_pet)}: Print results

\item \code{export(sc_pet)}: Export results as html table (see \code{\link[=export]{export()}})

}}
\examples{

## Calculate the PET and use a 99\%-CI for the additional calculation
# create random example data
design <- design(n = 5, slope = 0.2)
dat <- random_scdf(design, seed = 23)
pet(dat, ci = .99)

}
\seealso{
Other overlap functions:
\code{\link[=cdc]{cdc()}},
\code{\link[=ird]{ird()}},
\code{\link[=nap]{nap()}},
\code{\link[=overlap]{overlap()}},
\code{\link[=pand]{pand()}},
\code{\link[=pem]{pem()}},
\code{\link[=pnd]{pnd()}},
\code{\link[=select_phases]{select_phases()}},
\code{\link[=tau_u]{tau_u()}}
}
\author{
Juergen Wilbert
}
\concept{overlap functions}
