% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\description{
Several functions in scan are designed to work with pipes \code{\%>\%}.
This pipe function is directly imported from magrittr.
}
\details{
Since R 4.1 a pipe operator |> is included in base R. This pipe operator can be used along with scan perfectly fine.
}
