% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sc_rand.R
\name{plot.sc_rand}
\alias{plot.sc_rand}
\title{Plot random distirubtion}
\usage{
\method{plot}{sc_rand}(
  object,
  xlab = NA,
  ylab = "Frequency",
  title = "Random distribution",
  text_observed = "observed",
  color = "lightgrey",
  ...
)
}
\arguments{
\item{object}{Object returned from rand_test() function}

\item{...}{}
}
\description{
Plot random distirubtion
}
