% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rand.R
\name{plot_rand}
\alias{plot_rand}
\title{Plot random distribution}
\usage{
plot_rand(
  object,
  type = "hist",
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  text_observed = "observed",
  color = "lightgrey",
  ...
)
}
\arguments{
\item{object}{Object returned from the rand_test() function}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{title}{Plot title.}

\item{text_observed}{Text for marking the number of observed statistic.}

\item{color}{Bar color.}

\item{...}{Further arguments passed to the plot function.}
}
\description{
This function takes the return of the rand_test function and creates a
histogram with the distribution of the rand sample statistics.
}
