% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R, R/print-export-power.R
\name{power_test}
\alias{power_test}
\alias{print.sc_power}
\alias{export.sc_power}
\title{Empirical power analysis for single-case data}
\usage{
power_test(
  design,
  method = c("plm_level", "rand", "tauU"),
  effect = "level",
  n_sim = 100,
  design_is_one_study = TRUE,
  alpha_test = TRUE,
  power_test = TRUE,
  binom_test = FALSE,
  binom_test_alpha = FALSE,
  binom_test_power = FALSE,
  binom_test_correct = FALSE,
  ci = FALSE,
  alpha_level = 0.05
)

\method{print}{sc_power}(x, duration = FALSE, digits = 1, ...)

\method{export}{sc_power}(object, caption = NA, footnote = NA, filename = NA, round = 3, ...)
}
\arguments{
\item{design}{An object returned from the \code{design} function.}

\item{method}{A (named) list that defines the methods the power analysis is
based on. Each element can contain a function (that takes an scdf file and
returns a p-value) or a character string (the name of predefined
functions). default \code{method = list("plm_level", "rand", "tauU")} computes a
power analysis based on \code{\link[=tau_u]{tau_u()}}, \code{\link[=rand_test]{rand_test()}} and \code{\link[=plm]{plm()}} analyses.
(Further predefined functions are: "plm_slope", "plm_poisson_level",
"plm_poisson_slope", "hplm_level", "hplm_slope", "base_tau").}

\item{effect}{Either \code{"level"} or \code{"slope"}. The respective effect of the
provided design is set to 0 when computing the alpha-error proportion.}

\item{n_sim}{Number of sample studies created for the the Monte-Carlo study.
Default is \code{n = 100}. Ignored if design_is_one_study = FALSE.}

\item{design_is_one_study}{If TRUE, the design is assumed to define all cases
of one study that is repeatedly randomly created \code{n_sim} times. If FALSE,
one random case is generated for each design case provided in the design
object. This is useful for very specific complex simulation studies.
Default is TRUE.}

\item{alpha_test}{Logical. If TRUE, alpha error is calculated.}

\item{power_test}{Logical. If TRUE, power is calculated.}

\item{binom_test}{Shortcut. When set TRUE, binom_test_power is set to 0.80,
binom_test_alpha is set to 0.05, and binom_test_correct is set to 0.875.}

\item{binom_test_alpha}{Either FALSE or a value. If a value is provided, a
binomial test is calculated testing if the alpha error proportion is less
than the provided value.}

\item{binom_test_power}{Either FALSE or a value. If a value is provided, a
binomial test is calculated testing if the power is greater than the
provided value.}

\item{binom_test_correct}{Either FALSE or a value. If a value is provided, a
binomial test is calculated testing if the correct proportion is greater
than the provided value.}

\item{ci}{Either FALSE or a value. If a value is provided, confidence
intervals at the provided level are calculated for power, alpha error, and
correct proportions.}

\item{alpha_level}{Alpha level used to calculate the proportion of
significant tests. Default is \code{alpha_level = 0.05}.}

\item{x}{Object}

\item{duration}{If TRUE the duration for computation is printed.}

\item{digits}{The minimum number of significant digits to be use.
If set to "auto" (default), values are predefined.}

\item{...}{Further parameters passed to the print function}

\item{object}{An scdf or an object exported from a scan function.}

\item{caption}{Character string with table caption. If left NA (default) a
caption will be created based on the exported object.}

\item{footnote}{Character string with table footnote. Several strings are
combined into a footnote of several lines. If left NA (default) a footnote
will be created based on the exported object. \code{NULL} or \code{""} suppress the
footnote.}

\item{filename}{String containing the file name. If a filename is given the
output will be written to that file.}

\item{round}{Integer passed to the digits argument used to round values.}
}
\value{
An object of class \code{sc_power} with the power, alpha error, and
correct proportions for each provided method. If binomial tests are
requested, p values for these tests are also provided. If confidence
intervals are requested, these are also provided.
}
\description{
Conducts a Monte-Carlo study on the test-power and alpha-error probability of
a statistical function.
}
\details{
Based on a \code{\link[=design]{design()}} object, a large number of single-cases are generated
and re-analysed with a provided statistical function. The proportion of
significant analyses is the test power. In a second step, a specified effect
of the design object (specified in the \code{effect} parameter) is set to 0 and
again single-cases are generated and re-analysed. The proportion of
significant analyses is the alpha error probability.
}
\section{Functions}{
\itemize{
\item \code{print(sc_power)}: Print results

\item \code{export(sc_power)}: Export results as html table (see \code{\link[=export]{export()}})

}}
\examples{

## Assume you want to conduct a single-case study with 15 measurements
## (phases: A = 6 and B = 9) using a highly reliable test and
## an expected level effect of d = 1.4.
## A (strong) trend effect is trend = 0.05. What is the power?
## (Note: n_sims is set to 10. Set n_sims to 1000 for a serious calculation.)
design <- design(
  n = 1, phase_design = list(A = 6, B = 9),
  rtt = 0.8, level = 1.4, trend = 0.05
)
power_test(design, n_sim = 10)

## Would you achieve higher power by setting up a MBD with three cases?
design <- design(
  n = 3, phase_design = list(A = 6, B = 9),
  rtt = 0.8, level = 1.4, trend = 0.05
)
power_test(design, n_sim=10, method=list("hplm_level", "rand", "tauU_meta"))
}
\seealso{
\code{\link[=random_scdf]{random_scdf()}}, \code{\link[=design]{design()}}
}
\author{
Juergen Wilbert
}
