% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sc.R, R/print.sc_design.R,
%   R/print.sc_rci.R
\name{print.sc}
\alias{print.sc}
\alias{print.sc_design}
\alias{print.sc_rci}
\title{Print methods for scan objects}
\usage{
\method{print}{sc_design}(x, ...)

\method{print}{sc_rci}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Further parameters passed to the print function}

\item{digits}{The minimum number of significant digits to be use.
If set to "auto" (default), values are predefined.}
}
\description{
Print methods for scan objects
}
\keyword{internal}
