% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSC.R
\name{random}
\alias{random}
\alias{rSC}
\alias{design_rSC}
\title{Single-case data generator}
\usage{
rSC(design = NULL, round = NA, random.names = FALSE, seed = NULL, ...)

design_rSC(
  n = 1,
  phase.design = list(A = 5, B = 15),
  trend = 0,
  level = list(0),
  slope = list(0),
  rtt = 0.8,
  m = NULL,
  s = 10,
  start_value = 50,
  extreme.p = list(0),
  extreme.d = c(-4, -3),
  missing.p = 0,
  distribution = "normal",
  n_trials = NULL,
  MT = NULL,
  B.start = NULL
)
}
\arguments{
\item{design}{A design matrix which is created by design_rSC and specifies
all parameters.}

\item{round}{Rounds the scores to the defined decimal. To round to the second
decimal, set \code{round = 2}.}

\item{random.names}{Is \code{FALSE} by default. If set \code{random.names =
TRUE} cases are assigned random first names. If set \code{"neutral", 
"male" or "female"} only gender neutral, male, or female names are chosen. 
The names are drawn from the 2,000 most popular names for newborns in 2012 
in the U.S. (1,000 male and 1,000 female names).}

\item{seed}{A seed number for the random generator.}

\item{...}{arguments that are directly passed to the design_rSC function 
for a more concise coding.}

\item{n}{Number of cases to be designed (Default is \code{n = 1}).}

\item{phase.design}{A list defining the length and label of each phase.
E.g., \code{phase.length = list(A1 = 10, B1 = 10, A2 = 10, B2 = 10)}.
Use vectors if you want to define different values for each case \code{phase.length = list(A = c(10, 15), B = c(10, 15)}.}

\item{trend}{Defines the effect size of a trend added incrementally 
to each measurement across the whole data-set. To assign different trends
to several single-cases, use a vector of values 
(e.g. \code{trend = c(.1, .3, .5)}).
If the number of cases exceeds the length of the vector, values are
recycled. When using a 'gaussian' 
distribution, the \code{trend} parameters indicate effect size \emph{d} 
changes.  When using a binomial or poisson distribution, \code{trend}
indicates an increase in points / counts per measurement.}

\item{level}{A list that defines the level increase (effect size \emph{d}) 
at the beginning of each phase relative to the previous phase 
(e.g. \code{list(A = 0, B = 1)}). The first element must be zero as the 
first phase of a single-case has no level effect (if you have one less 
list element than the number of phases, scan will add a leading element 
with 0 values). Use vectors to define variable level effects for each case 
(e.g. \code{list(A = c(0, 0), B = c(1, 2)})). When using a 'gaussian' 
distribution, the \code{level} parameters indicate effect size \emph{d} 
changes. When using a binomial or poisson distribution, \code{level} 
indicates an increase in points / counts with the onset of each phase.}

\item{slope}{A list that defines the increase per measurement for each phase 
compared to the previous 
phase. \code{slope = list(A = 0, B = .1} generates an incremental increase 
of 0.1 per measurement starting at the B phase. The 
first list element must be zero as the first phase of a single-case has no 
slope effect (if you have one less list element than the number of phases, 
scan will add a leading element with 0 values). Use vectors to define 
variable slope effects for each case (e.g. \code{list(A = c(0, 0), 
B = c(0.1, 0.2)})). If the number of cases exceeds the length of the 
vector, values are recycled. When using a 'gaussian' distribution, the 
\code{slope} parameters indicate effect size \emph{d} changes per 
measurement. When using a binomial or poisson distribution, \code{slope} 
indicates an increase in points / counts per measurement.}

\item{rtt}{Reliability of the underlying simulated measurements. Set
\code{rtt = .8} by default. To assign different reliabilities to several
single-cases, use a vector of values (e.g. \code{rtt = c(.6, .7, .8)}). If
the number of cases exceeds the length of the vector, values are repeated.
\code{rtt} has no effect when you're using binomial or poisson 
distributions.}

\item{m}{Deprecated. Use start_value instead.}

\item{s}{Standard deviation used to calculate absolute values from level,
slope, trend effects and to calculate and error distribution from the 
\code{rtt} values. Set to \code{10} by default. 
To assign different variances to several single-cases, use a vector 
of values (e.g. \code{s = c(5, 10, 15)}). If the number of cases 
exceeds the length of the vector, values are recycled.
if the distribution is 'poisson' or 'binomial' s is not applied.}

\item{start_value}{Starting value at the first measurement. Default
is \code{50}. When \code{distribution = "poission"} the start_value 
represents frequency. When \code{distribution = "binomial"} start_value 
must range between 0 and 1 and they represent the probability of on event. 
To assign different start values to several single-cases, 
use a vector of values (e.g. \code{c(50, 42, 56)}). If the number of cases
exceeds the length of the vector, values are recycled.}

\item{extreme.p}{Probability of extreme values. \code{extreme.p = .05} gives
a five percent probability of an extreme value. A vector of values assigns
different probabilities to multiple cases. If the number of cases exceeds
the length of the vector, values are recycled.}

\item{extreme.d}{Range for extreme values, expressed as effect size \emph{d}.
\code{extreme.d = c(-7,-6)} uses extreme values within a range of -7 and -6
standard deviations. In case of a binomial or poisson distribution,
\code{extreme.d} indicates points / counts. Caution: the first value must
be smaller than the second, otherwise the procedure will fail.}

\item{missing.p}{Portion of missing values. \code{missing.p = 0.1} creates
10\% of all values as missing). A vector of values assigns different
probabilities to multiple cases. If the number of cases exceeds the length
of the vector, values are repeated.}

\item{distribution}{Erro distribution. Default is \code{"normal"}. 
Possible values are \code{"normal"}, \code{"binomial"}, 
and \code{"poisson"}.}

\item{n_trials}{If \code{distribution} (see below) is \code{"binomial"},
\code{n_trials} is the number of trials/observations/items.}

\item{MT}{Number of measurements (in each study). Default is \code{MT = 20}.}

\item{B.start}{Phase B starting point. The default setting \code{B.start = 6}
would assign the first five scores (of each case) to phase A, and all
following scores to phase B. To assign different starting points for a set
of multiple single-cases, use a vector of starting values (e.g.
\code{B.start = c(6, 7, 8)}). If the number of cases exceeds the length of
the vector, values will be recycled.}
}
\value{
A single-case data frame. See \code{\link{scdf}} to learn 
  about this format.
}
\description{
The \code{rSC} function generates random single-case data frames
for monte-carlo studies and demonstration purposes.
\code{design_rSC} is used to set up a design matrix with all parameters 
needed for the \code{rSC} function.
}
\examples{

## Create random single-case data and inspect it
design <- design_rSC(
  n = 3, rtt = 0.75, slope = 0.1, extreme.p = 0.1,
  missing.p = 0.1
)
dat <- rSC(design, round = 1, random.names = TRUE, seed = 123)
describeSC(dat)
plotSC(dat)

## And now have a look at poisson-distributed data
design <- design_rSC(
  n = 3, B.start = c(6, 10, 14), MT = c(12, 20, 22), start_value = 10,
  distribution = "poisson", level = -5, missing.p = 0.1
)
dat <- rSC(design, seed = 1234)
pand(dat, decreasing = TRUE, correction = FALSE)
}
\author{
Juergen Wibert
}
\keyword{datagen}
