% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-package.R
\docType{package}
\name{scan-package}
\alias{scan-package}
\alias{scan}
\title{Single-Case Data Analyses}
\description{
A collection of procedures for analysing, visualising, and managing
single-case data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jazznbass/scan/}
  \item \url{https://jazznbass.github.io/scan-Book/}
  \item \url{https://github.com/jazznbass/scan}
  \item Report bugs at \url{https://github.com/jazznbass/scan/issues}
}

}
\author{
Juergen Wilbert [aut, cre]
}
\keyword{internal}
\keyword{package}
