% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cases.R
\name{select_cases}
\alias{select_cases}
\title{Select a subset of cases from an scdf}
\usage{
select_cases(scdf, ...)
}
\arguments{
\item{scdf}{A single-case data frame. See \code{\link[=scdf]{scdf()}} to learn about this
format.}

\item{...}{Selection criteria. Either numeric, objectnames, or as characters.}
}
\value{
An scdf with a subset of cases.
}
\description{
This function allows to select a subset of cases from an scdf by
specifying either the case names or their numeric indices. Negative
selection is also supported.
}
\examples{
select_cases(exampleAB, Johanna, Karolina)
select_cases(exampleAB, c(Johanna, Karolina))
select_cases(exampleAB, 1,2)
select_cases(exampleAB, 1:2)
select_cases(exampleAB, -Johanna)
select_cases(exampleAB, -c(Johanna, Karolina))
v <- c("Moritz", "Jannis")
select_cases(exampleA1B1A2B2, v)
}
\seealso{
Other data manipulation functions:
\code{\link[=add_l2]{add_l2()}},
\code{\link[=as.data.frame.scdf]{as.data.frame.scdf()}},
\code{\link[=as_scdf]{as_scdf()}},
\code{\link[=batch_apply]{batch_apply()}},
\code{\link[=fill_missing]{fill_missing()}},
\code{\link[=moving_median]{moving_median()}},
\code{\link[=outlier]{outlier()}},
\code{\link[=ranks]{ranks()}},
\code{\link[=rescale]{rescale()}},
\code{\link[=scdf]{scdf()}},
\code{\link[=set_vars]{set_vars()}},
\code{\link[=shift]{shift()}},
\code{\link[=smooth_cases]{smooth_cases()}},
\code{\link[=standardize]{standardize()}},
\code{\link[=truncate_phase]{truncate_phase()}}
}
\author{
Juergen Wilbert
}
\concept{data manipulation functions}
