% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf.R
\name{setvarSCDF<-}
\alias{setvarSCDF<-}
\alias{setvarSCDF}
\title{Set the variables of an scdf}
\usage{
setvarSCDF(x, var = "values") <- value
}
\arguments{
\item{x}{An scdf.}

\item{var}{A character string specifying which variable you like to assign ("values", "phase", or "mt").}

\item{value}{A character string with the variable name.}
}
\description{
By default scan uses the variable names 'values'
as the dependend variable, 'phase' as indicator of a phase, and 'mt' as the
measurement-time. These variable names are attributes of an scdf. But you
might have other variables you want to use (e.g., another dependend
variable). With this function you can change these attributes of an scdf.
}
\examples{
setvarSCDF(exampleAB_add) <- "depression"
describeSC(exampleAB_add)
setvarSCDF(exampleAB_add) <- "wellbeing"
}
