% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Show scdf cases in viewer}
\usage{
show(scdf, max = 5)
}
\arguments{
\item{scdf}{A scdf object}

\item{max}{Maximum number of cases opened}
}
\description{
Each case with in scdf opens a separate viewer tab.
}
\examples{
## show(exampleABC)
}
