% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message-system.R
\name{warn}
\alias{warn}
\title{Throws a warning
This is a convenience wrapper for internal use.}
\usage{
warn(...)
}
\arguments{
\item{...}{The message. This can be a character string or a
combination of character strings and variables.}
}
\description{
Throws a warning
This is a convenience wrapper for internal use.
}
\keyword{internal}
