% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_l2.R
\name{add_l2}
\alias{add_l2}
\title{Add level 2}
\usage{
add_l2(data, data_l2, cvar = "case")
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{data_l2}{A level 2 dataset.}

\item{cvar}{Character string with the name of the "case" variable in the 
L2 dataset (default is 'case').}
}
\value{
An scdf
}
\description{
Add level 2 variables from a level 2 dataset to an scdf file
}
\examples{
Leidig2018 \%>\% add_l2(Leidig2018_l2)
}
