% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_scdf.R
\name{as_scdf}
\alias{as_scdf}
\title{as_scdf
Converts a data frame to an scdf object.}
\usage{
as_scdf(
  object,
  cvar = "case",
  pvar = "phase",
  dvar = "values",
  mvar = "mt",
  phase_names = NULL,
  sort_cases = FALSE
)
}
\arguments{
\item{object}{A data.frame}

\item{cvar}{Sets the variable name of the "case" variable. Defaults to
\code{"case"}.}

\item{pvar}{Sets the variable name of the "phase" variable. Defaults to
\code{"phase"}.}

\item{dvar}{Sets the variable name of the "values" variable. Defaults to
\code{"values"}.}

\item{mvar}{Sets the variable name of the "mt" variable. Defaults to
\code{"mt"}.}

\item{phase_names}{A character vector with phase names. Defaults to the phase
names provided in the phase variable.}

\item{sort_cases}{If set TRUE, the resulting list is sorted by label names
(alphabetically increasing).}
}
\description{
as_scdf
Converts a data frame to an scdf object.
}
