% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf-functions.R
\name{as_scdf}
\alias{as_scdf}
\title{as_scdf
Converts a data frame to an scdf object}
\usage{
as_scdf(object)
}
\arguments{
\item{object}{A scdf object}
}
\description{
as_scdf
Converts a data frame to an scdf object
}
