% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrSC.R
\name{autocorrSC}
\alias{autocorrSC}
\title{Autocorrelation for single-case data}
\usage{
autocorrSC(data, dvar, pvar, mvar, lag.max = 3, ...)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{dvar}{Character string with the name of the independend variable.}

\item{pvar}{Character string with the name of the phase variable.}

\item{mvar}{Character string with the name of the measurement time variable.}

\item{lag.max}{The lag up to which autocorrelations will be computed.
Default is \code{lag.max = 3}.}

\item{...}{Further arguments passed to the \code{\link{acf}} function}
}
\value{
autocorr A data frame containing separate autocorrelations for each
phase and for all phases (for each single-case). If \code{lag.max} exceeds
the length of a phase minus one, NA is returned for this cell.
}
\description{
The autocorrSC function calculates autocorrelations within each phase and
across all phases.
}
\examples{

 
## Compute autocorrelations for a list of four single-cases with max.lag = 2
autocorrSC(Huber2014, lag.max = 2)

}
\seealso{
\code{\link{trendSC}}, \code{\link{plm}}, \code{\link{acf}}
}
\author{
Juergen Wilbert
}
\concept{Autocorrelation}
\concept{Seiral correlation}
