% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_apply.R
\name{batch_apply}
\alias{batch_apply}
\title{Apply a function to each element in an scdf.}
\usage{
batch_apply(scdf, fn, simplify = FALSE)
}
\arguments{
\item{scdf}{A list of inputs to apply the function to.}

\item{fn}{The function to apply to each element. Use a \code{.} as a
placeholder for the scdf (e.g. \code{describe(.)}).}

\item{simplify}{If simplify is TRUE and \code{fn} returns a vector of values,
\code{batch_apply} will return a data frame case names.}
}
\value{
A list of the output of each function call.
}
\description{
This function applies a given function to each case of a multiple case scdf,
returning a list of the output of each function call.
}
\examples{
batch_apply(exampleAB, coef(plm(.)))

}
