% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc.R
\name{cdc}
\alias{cdc}
\title{Conservative Dual-Criterion Method}
\usage{
cdc(
  data,
  dvar,
  pvar,
  mvar,
  decreasing = FALSE,
  trend.method = "OLS",
  phases = c(1, 2)
)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{dvar}{Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.}

\item{pvar}{Character string with the name of the phase variable. Defaults to the attributes in the scdf file.}

\item{mvar}{Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.}

\item{decreasing}{If you expect data to be lower in the B phase, set
\code{decreasing = TRUE}. Default is \code{decreasing = FALSE}.}

\item{trend.method}{Method used to calculate the trend line. Default is \code{trend.method = "OLS"}.
Possible values are: \code{"OLS"}, \code{"bisplit"}, and \code{"trisplit"}.}

\item{phases}{A vector of two characters or numbers indicating the two
phases that should be compared. E.g., \code{phases = c("A","C")} or
\code{phases = c(2,4)} for comparing the second to the fourth phase. Phases
could be combined by providing a list with two elements. E.g., \code{phases
= list(A = c(1,3), B = c(2,4))} will compare phases 1 and 3 (as A) against 2
and 4 (as B). Default is \code{phases = c("A","B")}.}
}
\value{
\item{cdc}{CDC Evaluation based on a p-value below .05.}
\item{cdc.exc}{Number of phase B datapoints indicating expected change.}
\item{cdc.nb}{Number of phase B datapoints.} \item{cdc.p}{P value of Binomial
Test.} \item{N}{Number of cases.} \item{decreasing}{Logical argument from
function call (see \code{Arguments} above).} \item{case.names}{Assigned name
of single-case.} \item{phases}{-}
}
\description{
The \code{cdc} function applies the Conservative Dual-Criterion Method
(Swoboda, Kratochwill, & Levin, 2010) to scdf objects. It compares phase B
data points to both phase A median and trend (OLS, bi-split, tri-split) and
with an additional .25 SD. A binomial test against a 50/50 distribution is
computed and p-values below .05 are labeled "systematic change".
}
\examples{

## Apply the CDC method (standard OLS line)
design <- design_rSC(n = 1, slope = 0.2)
dat <- rSC(design, seed = 42)
cdc(dat)

## Apply the CDC with Koenig's bi-split and an expected decrease in phase B.
cdc(exampleAB_decreasing, decreasing = TRUE, trend.method = "bisplit")

## Apply the CDC with Tukey's tri-split, comparing the first and fourth phase.
cdc(exampleABAB, trend.method = "trisplit", phases = c(1,4))


}
\author{
Timo Lueke
}
