% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Covert
Converts an scdf object to a text syntax}
\usage{
convert(scdf, file = "", study_name = "study")
}
\arguments{
\item{scdf}{A single-case data frame. See \code{\link{scdf}} to learn about
this format.}

\item{file}{A filename for exporting the syntax.}

\item{study_name}{Character string. Name of the single case study name.}
}
\description{
Covert
Converts an scdf object to a text syntax
}
\examples{
filename <- tempfile()
convert(exampleABC, file = filename)
source(filename)
all.equal(study, exampleABC)
unlink(filename)
}
