% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/random_scdf.R, R/read_scdf.R,
%   R/write_scdf.R
\name{deprecated-functions}
\alias{deprecated-functions}
\alias{design_rSC}
\alias{longSCDF}
\alias{corrected_tauSC}
\alias{trendSC}
\alias{truncateSC}
\alias{style_plotSC}
\alias{style.plotSC}
\alias{smoothSC}
\alias{shiftSC}
\alias{tauUSC}
\alias{describeSC}
\alias{rankSC}
\alias{power_testSC}
\alias{fillmissingSC}
\alias{overlapSC}
\alias{randSC}
\alias{rand.test}
\alias{autocorrSC}
\alias{rciSC}
\alias{outlierSC}
\alias{rSC}
\alias{readSC.excel}
\alias{readSC}
\alias{writeSC}
\title{List of old deprecated function names}
\usage{
design_rSC(...)

longSCDF(...)

corrected_tauSC(...)

trendSC(...)

truncateSC(...)

style_plotSC(...)

style.plotSC(...)

smoothSC(...)

shiftSC(...)

tauUSC(...)

describeSC(...)

rankSC(...)

power_testSC(...)

fillmissingSC(...)

overlapSC(...)

randSC(...)

rand.test(...)

autocorrSC(...)

rciSC(...)

outlierSC(...)

rSC(...)

readSC.excel(...)

readSC(...)

writeSC(...)
}
\arguments{
\item{...}{Arguments passed through to the new function.}
}
\description{
This is a list with functions names that have been replaced by new names
which are in line with R syntax guidelines. The old function names still
work. They are wrappers that call the new function.
}
\keyword{internal}
