% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.R, R/deprecated.R, R/fill_missing.R,
%   R/outlier.R, R/overlap.R, R/power_test.R, R/rand_test.R, R/random_scdf.R,
%   R/ranks.R, R/rci.R
\name{autocorrSC}
\alias{autocorrSC}
\alias{deprecated-functions}
\alias{design_rSC}
\alias{longSCDF}
\alias{corrected_tauSC}
\alias{trendSC}
\alias{truncateSC}
\alias{style_plotSC}
\alias{style.plotSC}
\alias{smoothSC}
\alias{shiftSC}
\alias{tauUSC}
\alias{describeSC}
\alias{fillmissingSC}
\alias{outlierSC}
\alias{overlapSC}
\alias{power_testSC}
\alias{randSC}
\alias{rand.test}
\alias{rSC}
\alias{rankSC}
\alias{rciSC}
\alias{rCi}
\title{List of old deprecated function names}
\usage{
autocorrSC(...)

design_rSC(...)

longSCDF(...)

corrected_tauSC(...)

trendSC(...)

truncateSC(...)

style_plotSC(...)

style.plotSC(...)

smoothSC(...)

shiftSC(...)

tauUSC(...)

describeSC(...)

fillmissingSC(...)

outlierSC(...)

overlapSC(...)

power_testSC(...)

randSC(...)

rand.test(...)

rSC(...)

rankSC(...)

rciSC(...)

rCi(...)
}
\arguments{
\item{...}{Arguments passed to the design function.}
}
\description{
This is a list with functions names that have been replaced by new names
which are in line with re R syntax guidelines. The old function names still 
work. They are wrappers that call the new function.
}
