% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/read_scdf.R, R/write_scdf.R
\name{deprecated-functions}
\alias{deprecated-functions}
\alias{tauUSC}
\alias{power_testSC}
\alias{fillmissingSC}
\alias{overlapSC}
\alias{randSC}
\alias{rand.test}
\alias{rciSC}
\alias{rSC}
\alias{readSC.excel}
\alias{readSC}
\alias{writeSC}
\title{List of old deprecated function names}
\usage{
tauUSC(...)

power_testSC(...)

fillmissingSC(...)

overlapSC(...)

randSC(...)

rand.test(...)

rciSC(...)

rSC(...)

readSC.excel(...)

readSC(...)

writeSC(...)
}
\arguments{
\item{...}{Arguments passed through to the new function.}
}
\description{
This is a list with functions names that have been replaced by new names
which are in line with R syntax guidelines. The old function names still
work. They are wrappers that call the new function.
}
\keyword{internal}
