% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describeSC}
\title{Descriptive statistics for single-case data}
\usage{
describe(data, dvar, pvar, mvar)

describeSC(...)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{dvar}{Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.}

\item{pvar}{Character string with the name of the phase variable. Defaults to the attributes in the scdf file.}

\item{mvar}{Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.}

\item{...}{Further arguments passed to the function.}
}
\value{
A data frame of descriptive statistics (for each single-case), i.e.:
number of observations, number of missing values, measures of central
tendency, variation, and trend.
}
\description{
The \code{describe} function provides common descriptive statistics for
single-case data.
}
\examples{


## Descriptive statistics for a study of three single-cases
describe(Grosche2011)

## Descriptives of a three phase design
describe(exampleABC)

\dontrun{
## Write descriptive statistics to .csv-file
study <- describe(Waddell2011)
write.csv(study$descriptives, file = "descriptives.csv")
}
}
\seealso{
\code{\link{overlap}}, \code{\link{plot.scdf}}
}
\author{
Juergen Wilbert
}
