% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-Data.R
\docType{data}
\name{example_data_sets}
\alias{example_data_sets}
\alias{Beretvas2008}
\alias{Borckardt2014}
\alias{Huitema2000}
\alias{Waddell2011}
\alias{byHeart2011}
\alias{Grosche2011}
\alias{Grosche2014}
\alias{GruenkeWilbert2014}
\alias{Huber2014}
\alias{example_A24}
\alias{exampleAB_simple}
\alias{exampleAB}
\alias{exampleAB_add}
\alias{exampleAB_decreasing}
\alias{exampleAB_score}
\alias{exampleABC}
\alias{exampleABC_outlier}
\alias{exampleABC_50}
\alias{exampleABAB}
\alias{exampleA1B1A2B2}
\alias{exampleA1B1A2B2_zvt}
\alias{exampleAB_50}
\alias{exampleAB_50.l2}
\alias{exampleABC_150}
\alias{exampleAB_mpd}
\alias{Lenz2013}
\alias{Leidig2018}
\alias{Leidig2018_l2}
\alias{SSDforR2017}
\alias{Parker2011}
\alias{Parker2007}
\alias{Parker2009}
\alias{Parker2009b}
\alias{Parker2011b}
\alias{Tarlow2017}
\title{Single-case example data sets}
\usage{
data(example_data_sets, package = "scan")
}
\description{
The \link{scan} package comes with a set of fictitious and authentic
single-case study data.
These data sets can be used to practice single-case data analysis
and to reproduce results from the respective publications.
\itemize{
\item \strong{Beretvas2008} — Fictitious single-case intervention study from Beretvas & Chung, 2008.
\item \strong{Borckardt2014} — Fictitious daily pain ratings evaluating a psychological treatment from Borckardt & Nash, 2014.
\item \strong{byHeart2011} — Multiple-baseline (11 cases) flash card vocabulary learning (Wilbert, unpublished).
\item \strong{example_A24} — Number of injuries on a German autobahn before and after implementation of a speedlimit (130km/h) (Ministerium fuer Infrastruktur und Landesplanung. Land Brandenburg).
\item \strong{exampleA1B1A2B2} — Fictitious A1-B1-A2-B2 example dataset.
\item \strong{exampleA1B1A2B2_zvt} — Non-fictitious A1-B1-A2-B2 example with ZVT (intelligence measure) and D2 (concentration measure) scores.
\item \strong{exampleAB} — Fictitious AB example dataset with three cases.
\item \strong{exampleAB_50} — Fictitious AB example dataset (50 cases)s.
\item \strong{exampleAB_50.l2} — Level-2 data for \strong{exampleAB_50}.
\item \strong{exampleAB_add} — Fictitious AB example with added covariates.
\item \strong{exampleAB_decreasing} — Fictitious AB example with an expected decreasing effect in the intervention phase.
\item \strong{exampleAB_mpd} — Fictitious example with different phase structures for each case.
\item \strong{exampleAB_score} — Fictitious AB example with binomial distributed score outcome.
\item \strong{exampleAB_simple} — Simple fictitious AB example with three cases.
\item \strong{exampleABAB} — Fictitious ABAB reversal design example.
\item \strong{exampleABC} — Fictitious ABC example dataset.
\item \strong{exampleABC_150} — Fictitious ABC example (150 cases).
\item \strong{exampleABC_50} — Fictitious ABC example (50 cases).
\item \strong{exampleABC_outlier} — Fictitious ABC example with outlier.
\item \strong{Grosche2011} — Multiple-baseline (three cases) from a direct-instructive reading intervention (Grosche, 2011).
\item \strong{Grosche2014} — Multiple-baseline (3×3 materials) reading intervention (Grosche, Lueke, & Wilbert, unpublished).
\item \strong{GruenkeWilbert2014} — Multiple-baseline (six cases) from a story mapping reading intervention (Gruenke, Wilbert, & Stegemann-Calder, 2013).
\item \strong{Huber2014} — Multiple-baseline (four cases) with DBR ratings from a behavioral compliance intervention (Huber, unpublished).
\item \strong{Huitema2000} — Fictitious single-case intervention study (Huitema & McKean, 2000).
\item \strong{Lenz2013} — Fictitious example (Lenz, 2013).
\item \strong{Leidig2018} — Multiple-baseline good behavior game intervention (Leidig et al., 2022).
\item \strong{Leidig2018_l2} — Level-2 data for \strong{Leidig2018} (Leidig et al., 2022).
\item \strong{Parker2007} — Example dataset after Parker et al. (2007).
\item \strong{Parker2009} — Example dataset after Parker et al. (2009).
\item \strong{Parker2009b} — Example dataset after Parker & Vannest (2009).
\item \strong{Parker2011} — Example dataset after Parker et al. (2011).
\item \strong{Parker2011b} — Example from Parker, Vannest, & Davis (2011).
\item \strong{SSDforR2017} — Example from the R package SSDforR.
\item \strong{Tarlow2017} — Fictitious single-case intervention study (Tarlow, 2017).
\item \strong{Waddell2011} — Fictitious single-case intervention study (Waddell, Nassar, & Gustafson, 2011).
}
}
\references{
Beretvas, S., & Chung, H. (2008). An evaluation of modified R2-change effect size indices for single-subject experimental designs. \emph{Evidence-Based Communication Assessment and Intervention, 2}, 120–128.

Borckardt, J. J., & Nash, M. R. (2014). Simulation modelling analysis for small sets of single-subject data collected over time. \emph{Neuropsychological Rehabilitation, 24}, 492–506.

Gruenke, M., Wilbert, J., & Stegemann-Calder, K. (2013). Analyzing the effects of story mapping on the reading comprehension of children with low intellectual abilities. \emph{Learning Disabilities: A Contemporary Journal, 11}, 51–64.

Grosche, M. (2011). Effekte einer direkt-instruktiven Foerderung der Lesegenauigkeit. \emph{Empirische Sonderpaedagogik, 3}, 147–161.

Huitema, B. E., & McKean, J. W. (2000). Design specification issues in time-series intervention models. \emph{Educational and Psychological Measurement, 60}, 38–58.

Lenz, A. S. (2013). Calculating Effect Size in Single-Case Research: A Comparison of Nonoverlap Methods. \emph{Measurement and Evaluation in Counseling and Development, 46}(1), 64–73.

Leidig, T., Casale, G., Wilbert, J., Hennemann, T., Volpe, R. J., Briesch, A., & Grosche, M. (2022). Individual, generalized, and moderated effects of the good behavior game on at-risk primary school students: A multilevel multiple baseline study using behavioral progress monitoring. \emph{Frontiers in Education, 7}. https://www.frontiersin.org/articles/10.3389/feduc.2022.917138

Parker, R. I., Hagan-Burke, S., & Vannest, K. (2007). Percentage of All Non-Overlapping Data (PAND) An Alternative to PND. The Journal of Special Education, 40(4), 194-204.

Parker, R. I., Vannest, K. J., & Brown, L. (2009). The improvement rate difference for single-case research. Exceptional Children, 75(2), 135-150.

Parker, R. I., & Vannest, K. (2009). An improved effect size for single-case research: Nonoverlap of all pairs. Behavior Therapy, 40(4), 357-367.

Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. (2011). Combining Nonoverlap and Trend for Single-Case Research: Tau-U. \emph{Behavior Therapy, 42}(2), 284–299. https://doi.org/10.1016/j.beth.2010.08.006

Parker, R. I., Vannest, K. J., & Davis, J. L. (2011). Effect Size in Single-Case Research: A Review of Nine Nonoverlap Techniques. Behavior Modification, 35(4), 303-322. https://doi.org/10.1177/0145445511399147

Tarlow, K. R. (2017). An Improved Rank Correlation Effect Size Statistic for Single-Case Designs: Baseline Corrected Tau. \emph{Behavior Modification, 41}(4), 427–467. https://doi.org/10.1177/0145445516676750

Waddell, D. E., Nassar, S. L., & Gustafson, S. A. (2011). Single-Case Design in Psychophysiological Research: Part II: Statistical Analytic Approaches. \emph{Journal of Neurotherapy, 15}, 160–169.
}
\author{
Juergen Wilbert
}
\keyword{datasets}
