\name{example-data}
\alias{Beretvas2008}
\alias{example1}
\alias{Grosche2011}
\alias{byHeart2011}
\alias{GruenkeWilbert2014}
\alias{Grosche2014}
\alias{Huber2014}
\alias{Beretvas2008}
\alias{Huitema2000}
\alias{Waddell2011}
\alias{Borckardt2014}

\title{Single-case data examples}
\description{
This is a list of example single-case data sets implemented in the package. 
}
########################add citations##################################
\details{
example1

Grosche2011

byHeart2011

GruenkeWilbert2014

Grosche2014

Huber2014

Beretvas2008

Huitema2000

Waddell2011

Borckardt2014

}
\author{
Juergen Wilbert
}

