% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R, R/export.sc_desc.R,
%   R/export.sc_nap.R, R/export.sc_overlap.R, R/export.sc_pem.R,
%   R/export.sc_pet.R, R/export.sc_pnd.R, R/export.sc_power.R,
%   R/export.sc_smd.R, R/export.sc_trend.R, R/export.scdf.R,
%   R/export.scdf_summary.R
\name{export}
\alias{export}
\alias{export.sc_desc}
\alias{export.sc_nap}
\alias{export.sc_overlap}
\alias{export.sc_pem}
\alias{export.sc_pet}
\alias{export.sc_pnd}
\alias{export.sc_power}
\alias{export.sc_smd}
\alias{export.sc_trend}
\alias{export.scdf}
\alias{export.scdf_summary}
\title{Export scan objects to html or latex}
\usage{
export(object, ...)

\method{export}{sc_desc}(
  object,
  caption = NA,
  footnote = NA,
  filename = NA,
  flip = FALSE,
  round = 2,
  ...
)

\method{export}{sc_nap}(
  object,
  caption = NA,
  footnote = NA,
  filename = NA,
  select = c("Case", "NAP", "NAP Rescaled", "w", "p", "d", "R²"),
  round = 2,
  ...
)

\method{export}{sc_overlap}(
  object,
  caption = NA,
  footnote = NULL,
  filename = NA,
  round = 2,
  decimals = 2,
  flip = FALSE,
  ...
)

\method{export}{sc_pem}(object, caption = NA, footnote = NA, filename = NA, round = 2, ...)

\method{export}{sc_pet}(object, caption = NA, footnote = NA, filename = NA, round = 1, ...)

\method{export}{sc_pnd}(
  object,
  caption = NA,
  footnote = NA,
  filename = NA,
  select = c("Case", "PND", "Total", "Exceeds"),
  round = 2,
  ...
)

\method{export}{sc_power}(object, caption = NA, footnote = NA, filename = NA, round = 3, ...)

\method{export}{sc_smd}(
  object,
  caption = NA,
  footnote = NA,
  filename = NA,
  select = c("Case", `Mean A` = "mA", `Mean B` = "mB", `SD A` = "sdA", `SD B` = "sdB",
    `SD Cohen` = "sd cohen", `SD Hedges` = "sd hedges", "Glass' delta", "Hedges' g",
    "Hedges' g correction", "Hedges' g durlak correction", "Cohen's d"),
  round = 2,
  decimals = 2,
  flip = FALSE,
  ...
)

\method{export}{sc_trend}(
  object,
  caption = NA,
  footnote = NA,
  filename = NA,
  round = 3,
  decimals = NULL,
  ...
)

\method{export}{scdf}(
  object,
  summary = FALSE,
  caption = NA,
  footnote = NA,
  filename = NA,
  cols,
  round = 2,
  ...
)

\method{export}{scdf_summary}(object, caption = NA, footnote = NA, filename = NA, round = 2, ...)
}
\arguments{
\item{object}{An scdf or an object exported from a scan function.}

\item{...}{Further Arguments passed to internal functions.}

\item{caption}{Character string with table caption. If left NA (default) a
caption will be created based on the exported object.}

\item{footnote}{Character string with table footnote. Several strings are
combined into a footnote of several lines. If left NA (default) a footnote
will be created based on the exported object. \code{NULL} or \code{""} suppress the
footnote.}

\item{filename}{String containing the file name. If a filename is given the
output will be written to that file.}

\item{flip}{If TRUE, some objects are exported with rows and columns flipped.}

\item{round}{Integer passed to the digits argument used to round values.}

\item{select}{A character vector containing the names of the variables to be
included. If the vector is named, the variables will be renamed
accordingly.}

\item{decimals}{Decimal places that are reported.}

\item{summary}{If TRUE, exports the summary of an \code{scdf}.}

\item{cols}{Defines which columns are included when exporting an scdf. It is
either a vector of variable names or the string "main" will select the
central variables.}
}
\value{
Returns or displays a specially formatted html (or latex) file.
}
\description{
Export creates html files of tables or displays them directly in the viewer
pane of rstudio. When applied in rmarkdown/quarto, tables can also be created
for pdf/latex output.
}
\details{
The function uses either the gt or the kableExtra package to create the
tables. The default engine is gt; in a document knitted to latex, kableExtra
is used instead, because gt tables receive neither a number nor a label
there, and for Word output gt is used even when kableExtra is set, because
kableExtra can not write Word tables. The engine can be set via the option
\code{scan.export.engine}. Additional
options for kable and kable_styling can be set via the options
\code{scan.export.kable} and \code{scan.export.kable_styling}.
The default options can be viewed and modified via
\code{options("scan.export.kable")} and
\code{options("scan.export.kable_styling")}.
}
