% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export scan objects to html or latex}
\usage{
export(object, filename = NULL, kable_styling_options = list(),
  kable_options = list(), cols, flip = FALSE, note = TRUE, ...)
}
\arguments{
\item{object}{An scdf}

\item{filename}{Character string with the filename. If a filename is provided
the output will be written into this file.}

\item{kable_styling_options}{list with arguments passed to the kable_styling function.}

\item{kable_options}{list with arguments passed to the kable function.}

\item{cols}{Defines which columns are included when a scdf is exported. It is either a vector 
with variable names or the string "main" will select the central variables.}

\item{flip}{If TRUE, some objects are displayed with rows and columns flipped.}

\item{note}{If TRUE additional information are printed below the table.}

\item{...}{Further Arguments passed to internal functions.}
}
\value{
Returns a specif formated html.
}
\description{
This function is in an experimental status.
Export creates html files of tables or displayes them directly in the viewer pane of rstudio.
When applied in rmarkdown, tables can also be created for pdf/latex output.
}
