% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export scan output to csv Files}
\usage{
export(object, filename = NULL, sep = ",", dec = ".",
  type = "html", caption = "", flip = FALSE, kable.style = list(),
  cols, ...)
}
\arguments{
\item{object}{An scdf}

\item{filename}{Character string with the filename.}

\item{sep}{Sign for data seperator.}

\item{dec}{Sign for decimal seperator.}

\item{type}{...}

\item{caption}{...}

\item{flip}{...}

\item{kable.style}{= list with arguments passed to the kable_styling function,}

\item{cols}{...}

\item{...}{Further Arguments passed to the kable function.}
}
\value{
Returns a specif formated html.
}
\description{
Export scan output to csv Files
}
