% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetches elements from scan objects
Getter function for scan objects}
\usage{
fetch(object, what, ...)
}
\arguments{
\item{object}{Object returned from a scan function.}

\item{what}{Element/part to be extracted.}

\item{...}{Further parameters passed to the function.}
}
\value{
An object of the respective regression model class.
}
\description{
Fetches elements from scan objects
Getter function for scan objects
}
