% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplm.R
\name{hplm}
\alias{hplm}
\title{Hierarchical piecewise linear model / piecewise regression}
\usage{
hplm(data, model = "B&L-B", method = "ML", control = list(opt = "optim"),
  random.slopes = FALSE, lr.test = FALSE, ICC = TRUE, trend = TRUE,
  level = TRUE, slope = TRUE, fixed = NULL, random = NULL,
  update.fixed = NULL, data.l2 = NULL, ...)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about
this format.}

\item{model}{Model used for calculating the slope parameter (see Huitema &
McKean, 2000). Default is \code{model = "B&L-B"}. Possible values are:
\code{"B&L-B"}, \code{"H-M"}, \code{"Mohr#1"}, \code{"Mohr#2"}, \code{"JW"}, \code{"JW2"}, and
\code{"Manly"}.}

\item{method}{Method used to fit your model. Pass \code{"REML"} to maximize
the restricted log-likelihood or \code{"ML"} for maximized log-likelihood.
Default is \code{"ML"}.}

\item{control}{A list of settings for the estimation algorithm, replacing
the default values passed to the function \code{lmeControl} of the
\code{nlme} package.}

\item{random.slopes}{If \code{random.slopes = TRUE} random slope effects of
the level, trend, and treatment parameter are estimated.}

\item{ICC}{If \code{ICC = TRUE} an intraclass-correlation is estimated.}

\item{data.l2}{A dataframe providing additional variables at Level 2. The
scdf File has to have names for all cases and the Level 2 dataframe has to
have a column named 'cases' with the names of the cases the Level 2
variables belong to.}

\item{formula}{Defaults to the standard piecewise regression model. The
parameter 'inter' adresses the intercation term defined based on the method
provide in the model argument (e.g., "B&L-B"). The formula can be changed
for example to include further L1 or L2 variables into the regression model.}
}
\value{
\item{model}{List containing infromation about the applied model} 
\item{N}{Number of single-cases.}
\item{hplm}{Object of class lme contaning the multilevel model} 
\item{model.0}{Object of class lme containing the Zero Model.} 
\item{model.without}{Object of class gls containing fixed effect model.}
\item{ICC}{List containing intraclass correlation and test parameters.}
}
\description{
The \code{hplm} function computes a hierarchical piecewise regression model.
}
\examples{

## Compute hplm model on a MBD over fifty cases (restricted log-likelihood)
hplm(exampleAB_50, method = "REML", random.slopes = TRUE)

## Analyzing with additional L2 variables
hplm(exampleAB_50, data.l2 = exampleAB_50.l2, update.fixed = .~. + age + sex + age:sex)
}
\seealso{
\code{\link{plm}}
}
\author{
Juergen Wilbert
}
