\name{longSCDF}
\alias{longSCDF}
\title{
Create a single data frame of a list of single-case dataframes
}
\description{This functions helps you to create a single data frame from a list of multiple single cases.
}
\usage{
longSCDF(data)
}

\arguments{
  \item{data}{
A list with single-case dataframes
}
}
\value{
Returns a data frame with one or multiple single-cases divided by the variabele 'case'. 
}
\author{
Juergen Wilbert
}

