% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_makesingleSC.R
\name{makesingleSC}
\alias{makesingleSC}
\title{Aggregate multiple single-cases into one case}
\usage{
makesingleSC(data, scale = FALSE, type = "add")
}
\arguments{
\item{data}{A vector with measurements, a data frame or a list of data
frames.}

\item{scale}{Unused}

\item{type}{By default values with the same measurement are added. If type
is set to "mean" or "median", values of the same measurement are replaced
with their mean or median. Default is "add".}
}
\description{
The makesingleSC function combines multiple single-case data frames into one
single-case data frame.
}
\details{
The algorithm works the following way: \enumerate{ \item All values of each
single-case are centred with respect to each case's phase A mean.  \item The
phase A values of all single-cases are combined in ascending order of their
measurement times.  \item The phase B values of all single-cases are combined
in ascending order of their measurement times.  \item Phase B values are
appended to phase A values. The measurement times of phase B are shifted to
start with the next MT after the end of phase A. }
}
\examples{

##Function deprecated
## please do not use it!
}
\seealso{
\code{\link{scdf}}, \code{\link{longSCDF}},
\code{\link{writeSC}}
}
\author{
Juergen Wilbert
}
\keyword{manip}
