% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf-functions.R
\name{na.omit.scdf}
\alias{na.omit.scdf}
\title{scdf objects
Removes any row with a missing value}
\usage{
\method{na.omit}{scdf}(object, ...)
}
\arguments{
\item{object}{A scdf.}

\item{...}{not implemented yet.}
}
\value{
A scdf object.
}
\description{
scdf objects
Removes any row with a missing value
}
