% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{plm}
\alias{plm}
\title{Piecewise linear model / piecewise regression}
\usage{
plm(data, AR = 0, model = "B&L-B", phases = NULL, family = "gaussian",
  trend = TRUE, level = TRUE, slope = TRUE, formula = NULL,
  update = NULL, na.action = na.omit, ...)
}
\arguments{
\item{data}{A single-case data frame or a list of single-case data frames.
See \code{\link{makeSCDF}} to learn about this format.}

\item{AR}{Maximal lag of autoregression. Modeled based on the
Autoregressive-Moving Average (ARMA) function.  When AR is set, the family
argument must be set to \code{family = "gaussian"}.}

\item{model}{Regression model used for computation (see Huitema & McKean,
2000). Default is \code{model = "B&L-B"}. Possible values are:
\code{"B&L-B"}, \code{"H-M"}, \code{"Mohr#1"}, \code{"Mohr#2"},
\code{"Manly"}, \code{"JW"}, and , \code{"JW2"}.}

\item{phases}{-}

\item{family}{Set the distributioin family. Defaults to a gaussian
distribution. See the \code{family} function for more details.}

\item{formula}{-}

\item{na.action}{Defines how to deal with missing values}

\item{...}{Further arguments passed to the glm function.}
}
\value{
\item{model}{Character string from function call (see
\code{Arguments} above).} \item{F}{F value for specified model.}
\item{df1}{Degrees of freedom (Regression).} \item{df2}{Degrees of freedom
(Residual).} \item{p}{P value for specified model.} \item{R2}{Explained
variance R squared.} \item{R2.adj}{Adjusted R squared.}
\item{count.data}{Logical argument from function call (see \code{Arguments}
above).} \item{ES.slope}{Effect size / Explained variance gain of slope.}
\item{ES.trend}{Effect size / Explained variance gain of trend.}
\item{full.model}{Full regression model list (including \code{coefficients},
\code{residuals} and many others} \item{MT}{Number of measurements.}
\item{data}{Single-case data frame passed to the function.} \item{N}{Number
of single-cases.} \item{family}{Character string from function call (see
\code{Arguments} above).}
}
\description{
The \code{plm} function computes a piecewise regression model (see Huitema &
McKean, 2000).
}
\examples{

## Compute a piecewise regression model for a random single-case
set.seed(123)
AB <- design.rSC(n = 1, phase.design = list(A = 10, B = 20), 
            level = list(A = 0, B = 1), slope = list(A = 0, B = 0.05), 
            trend = list(0.05))
dat <- rSC(design = AB)
plm(dat, AR = 3)

## Another example with a more complex design

set.seed(123)
A1B1A2B2 <- design.rSC(n = 1, rtt = 0.8,   m = 50, s = 10,
                  phase.design = list(A1 = 15, B1 = 20, A2 = 15, B2 = 20), 
                  level = list(A1 = 0, B1 = 1, A2 = -1, B2 = 1),
                  slope = list(A1 = 0, B1 = 0.0, A1 = 0, B2 = 0.0),
                  trend = list(0.0))
dat <- rSC(design = A1B1A2B2)
plm(dat, model = "JW")

## no slope effects were found. Therefore you might want to drop slope estimation:
plm(dat, slope = FALSE, model = "JW")

## and now drop the trend estimation as well
plm(dat, slope = FALSE, trend = FALSE, model = "JW")





}
\references{
Beretvas, S., & Chung, H. (2008). An evaluation of modified
R2-change effect size indices for single-subject experimental designs.
\emph{Evidence-Based Communication Assessment and Intervention, 2}, 120-128.

Huitema, B. E., & McKean, J. W. (2000). Design specification issues in
time-series intervention models. \emph{Educational and Psychological
Measurement, 60}, 38-58.
}
\seealso{
\code{\link{hplm}}
}
\author{
Juergen Wilbert
}
