% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{power_test}
\alias{power_test}
\alias{power_testSC}
\title{Empirical power analysis for single-case data}
\usage{
power_test(
  design,
  stat = c("plm_level", "rand", "tauU"),
  effect = "level",
  n_sim = 100,
  design_is_one_study = TRUE,
  alpha_stat = TRUE,
  power_stat = TRUE,
  binom_test = 0.5,
  alpha_level = 0.05
)

power_testSC(...)
}
\arguments{
\item{design}{An object created by design_rSC}

\item{stat}{Defines the tests the power analysis is based on. The
default \code{stat = c("plm_level", "rand", "tauU")} computes a power 
analysis based on \code{\link{tau_u}}, \code{\link{randSC}} and 
\code{\link{plm}} analyses. Further possibilities are: "plm_slope", 
"plm_poisson_level", "plm_poisson_slope", "hplm_level", "hplm_slope", 
"base_tau".}

\item{effect}{Either "level" or "slope".}

\item{n_sim}{Number of sample studies created for the the Monte-Carlo study.
Default is \code{n = 100}. Ignored if design_is_one_study = FALSE.}

\item{design_is_one_study}{If TRUE, the design is assumed to define all 
cases of one study that is repeatedly randomly created \code{n_sim} times. 
If false, the design is assumed to contain all cases that from which a 
random sample is generated. This is useful for very specific complex 
simulation studies.}

\item{alpha_level}{Alpha level used to calculate the proportion of 
significant tests. Default is \code{alpha_level = 0.05}.}

\item{...}{Further arguments passed to the function.}
}
\description{
The \code{power_test} command conducts a Monte-Carlo study on the
test-power and alpha-error of a set of single-cases. The distribution 
values of the Monte-Carlo sample are either specified by the user or 
estimated based on actual data.
}
\examples{

## Assume you want to conduct a single-case study with 15 MTs, using a highly reliable test,
## an expected level effect of \eqn{d = 1.4}, and randomized start points between MTs 5
## and 12 can you expect to identify the effect using plm or randomization test?
design <- design_rSC(
  n = 1, phase.design = list(A = 6, B = 9), 
  rtt = 0.8, level = 1.4
)
res <- power_test(design, n_sim = 10)

## Would you achieve higher power by setting up a MBD with three cases?
design <- design_rSC(
  n = 3, phase.design = list(A = 6, B = 9), 
  rtt = 0.8, level = 1.4
)
res <- power_test(design, n_sim = 10, stat = c("hplm_level", "rand"))

}
\seealso{
\code{\link{plm}}, \code{\link{randSC}}
}
\author{
Juergen Wilbert
}
