% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sc.R
\name{print.sc}
\alias{print.sc}
\alias{print.sc_ac}
\alias{print.sc_cdc}
\title{Print method for scan objects}
\usage{
\method{print}{sc}(x, ...)

\method{print}{sc_ac}(x, ...)

\method{print}{sc_cdc}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Further parameters passed to the print function}
}
\description{
Print method for scan objects
}
