% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-export-plm.R, R/print.sc.R,
%   R/print.sc_ac.R, R/print.sc_bctau.R, R/print.sc_cdc.R, R/print.sc_desc.R,
%   R/print.sc_design.R, R/print.sc_nap.R, R/print.sc_outlier.R,
%   R/print.sc_overlap.R, R/print.sc_pem.R, R/print.sc_pet.R, R/print.sc_pnd.R,
%   R/print.sc_power.R, R/print.sc_rand.R, R/print.sc_rci.R, R/print.sc_smd.R,
%   R/print.sc_trend.R
\name{print.sc_plm}
\alias{print.sc_plm}
\alias{print.sc}
\alias{print.sc_ac}
\alias{print.sc_bctau}
\alias{print.sc_cdc}
\alias{print.sc_desc}
\alias{print.sc_design}
\alias{print.sc_nap}
\alias{print.sc_outlier}
\alias{print.sc_overlap}
\alias{print.sc_pem}
\alias{print.sc_pet}
\alias{print.sc_pnd}
\alias{print.sc_power}
\alias{print.sc_rand}
\alias{print.sc_rci}
\alias{print.sc_smd}
\alias{print.sc_trend}
\title{Print methods for scan objects}
\usage{
\method{print}{sc_plm}(x, lag_max = 3, ...)

\method{print}{sc_ac}(x, digits = "auto", ...)

\method{print}{sc_bctau}(x, nice = TRUE, digits = "auto", ...)

\method{print}{sc_cdc}(x, nice = TRUE, ...)

\method{print}{sc_desc}(x, digits = "auto", ...)

\method{print}{sc_design}(x, ...)

\method{print}{sc_nap}(x, digits = "auto", nice = TRUE, complete = FALSE, ...)

\method{print}{sc_outlier}(x, digits = "auto", ...)

\method{print}{sc_overlap}(x, digits = "auto", ...)

\method{print}{sc_pem}(x, ...)

\method{print}{sc_pet}(x, digits = 3, ...)

\method{print}{sc_pnd}(x, ...)

\method{print}{sc_power}(x, duration = FALSE, digits = 1, ...)

\method{print}{sc_rand}(x, ...)

\method{print}{sc_rci}(x, digits = 3, ...)

\method{print}{sc_smd}(x, digits = "auto", ...)

\method{print}{sc_trend}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object}

\item{lag_max}{Maximum lag to be reported for autocorrelation of residuals.
Default is \code{3}. Set \code{FALSE} for no report of autocorrelations.}

\item{...}{Further parameters passed to the print function}

\item{digits}{The minimum number of significant digits to be use.
If set to "auto" (default), values are predefined.}

\item{nice}{If set TRUE (default) output values are rounded and optimized for
publication tables.}

\item{duration}{If TRUE the duration for computation is printed.}
}
\description{
Print methods for scan objects
}
\keyword{internal}
