% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sc.R, R/print.sc_ac.R,
%   R/print.sc_bctau.R, R/print.sc_cdc.R, R/print.sc_desc.R,
%   R/print.sc_design.R, R/print.sc_hplm.R, R/print.sc_mplm.R, R/print.sc_nap.R,
%   R/print.sc_outlier.R, R/print.sc_overlap.R, R/print.sc_pand.R,
%   R/print.sc_pem.R, R/print.sc_pet.R, R/print.sc_plm.R, R/print.sc_pnd.R,
%   R/print.sc_power.R, R/print.sc_rand.R, R/print.sc_rci.R, R/print.sc_smd.R,
%   R/print.sc_tauu.R, R/print.sc_trend.R
\name{print.sc}
\alias{print.sc}
\alias{print.sc_ac}
\alias{print.sc_bctau}
\alias{print.sc_cdc}
\alias{print.sc_desc}
\alias{print.sc_design}
\alias{print.sc_hplm}
\alias{print.sc_mplm}
\alias{print.sc_nap}
\alias{print.sc_outlier}
\alias{print.sc_overlap}
\alias{print.sc_pand}
\alias{print.sc_pem}
\alias{print.sc_pet}
\alias{print.sc_plm}
\alias{print.sc_pnd}
\alias{print.sc_power}
\alias{print.sc_rand}
\alias{print.sc_rci}
\alias{print.sc_smd}
\alias{print.sc_tauu}
\alias{print.sc_trend}
\title{Print methods for scan objects}
\usage{
\method{print}{sc_ac}(x, digits = "auto", ...)

\method{print}{sc_bctau}(x, nice = TRUE, digits = "auto", ...)

\method{print}{sc_cdc}(x, nice = TRUE, ...)

\method{print}{sc_desc}(x, digits = "auto", ...)

\method{print}{sc_design}(x, ...)

\method{print}{sc_hplm}(x, ...)

\method{print}{sc_mplm}(x, digits = "auto", std = FALSE, ...)

\method{print}{sc_nap}(x, digits = "auto", ...)

\method{print}{sc_outlier}(x, digits = "auto", ...)

\method{print}{sc_overlap}(x, digits = "auto", ...)

\method{print}{sc_pand}(x, ...)

\method{print}{sc_pem}(x, ...)

\method{print}{sc_pet}(x, ...)

\method{print}{sc_plm}(x, ...)

\method{print}{sc_pnd}(x, ...)

\method{print}{sc_power}(x, duration = FALSE, digits = 3, ...)

\method{print}{sc_rand}(x, ...)

\method{print}{sc_rci}(x, ...)

\method{print}{sc_smd}(x, digits = "auto", ...)

\method{print}{sc_tauu}(x, complete = FALSE, digits = "auto", ...)

\method{print}{sc_trend}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object}

\item{digits}{The minimum number of significant digits to be use. 
If set to "auto" (default), values are predefined.}

\item{...}{Further parameters passed to the print function}

\item{nice}{If set TRUE (default) output values are rounded and optimized for
publication tables.}

\item{std}{If TRUE, a table with standardized estimates is included.}

\item{duration}{If TRUE the duration for computation is printed.}

\item{complete}{Print further parameters.}
}
\description{
Print methods for scan objects
}
