% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sc.R
\name{print.sc}
\alias{print.sc}
\alias{print.sc_ac}
\alias{print.sc_cdc}
\alias{print.sc_bctau}
\alias{print.sc_desc}
\alias{print.sc_design}
\alias{print.sc_hplm}
\title{Print method for scan objects}
\usage{
\method{print}{sc}(x, ...)

\method{print}{sc_ac}(x, ...)

\method{print}{sc_cdc}(x, nice = TRUE, ...)

\method{print}{sc_bctau}(x, nice = TRUE, ...)

\method{print}{sc_desc}(x, ...)

\method{print}{sc_design}(x, ...)

\method{print}{sc_hplm}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Further parameters passed to the print function}

\item{nice}{If set TRUE (default) output values are rounded and optimized for
publication tables.}
}
\description{
Print method for scan objects
}
