% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sc.R
\name{print.sc}
\alias{print.sc}
\title{Print method for return objects of scan}
\usage{
\method{print}{sc}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{-}
}
\description{
Print method for return objects of scan
}
