% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scdf.R
\name{print.scdf}
\alias{print.scdf}
\title{Print an scdf}
\usage{
\method{print}{scdf}(x, cases = 3, rows = 15, cols = "all",
  row.names = FALSE, long = FALSE, ...)
}
\arguments{
\item{x}{An scdf object}

\item{cases}{Number of cases to be printed.}

\item{rows}{Number of rows to be printed.}

\item{cols}{Columns to be printed.}

\item{row.names}{Logical. If TRUE row names are printed.}

\item{long}{Logical. If TRUE cases are printed in one by a time.}

\item{...}{Further arguments passed to the print function.}
}
\description{
Print an scdf
}
