% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSC.R
\name{rankSC}
\alias{rankSC}
\title{Rank-transformation of single-case data files}
\usage{
rankSC(data)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about
this format.}
}
\value{
An \code{scdf} object where the values of the \code{values} variable(s) are replaced with ranks.
}
\description{
Rank-transformation of single-case data files
}
\examples{
Huber2014_rank <- rankSC(Huber2014)
plot(Huber2014_rank, style = "grid2")
}
\author{
Juergen Wilbert
}
