% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSC.R
\name{readSC}
\alias{readSC}
\alias{readSC.excel}
\title{Read single-case data from files}
\usage{
readSC(filename = NULL, sep = ",", dec = ".", sort.labels = FALSE,
  type = "csv", ...)
}
\arguments{
\item{filename}{A character string defining the file to be imported (e.g.
\code{"SC_Anita.csv"}. If filename is left empty a dialog box for choosing
will be opened.}

\item{sep}{The field separator string. Values within rows have to be
separated by this string. Default is \code{sep = ","}.}

\item{dec}{The string used for decimal points in the file. Must be a single
character. Default is \code{dec = "."}}

\item{sort.labels}{If set TRUE, the resulting list is sorted by label names
(alphabetically increasing).}

\item{type}{Format of the file to be imported. Either "csv" or "excel" is
possible.}

\item{\dots}{Further arguments passed to the \code{\link{read.table}}
command.}
}
\value{
Returns a single-case data frame. See \code{\link{scdf}} to learn
about the format of these data frames.
}
\description{
Use the \code{readSC} function to import single-case data from structured
.csv or the \code{readSC.excel} function for importing excel files.
}
\examples{

## Read SC-data from a file named "study1.csv" in your working directory
# study1 <- readSC("study1.csv")

## Read SC-data from a .csv-file with semicolon as field and comma as decimal separator
# study2 <- readSC("study2.csv", sep = ";", dec = ",")

}
\seealso{
\code{\link{read.table}}, \code{\link{writeSC}}, \code{\link{scdf}}
}
\author{
Juergen Wilbert
}
\keyword{manip}
