% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_names.R
\name{sample_names}
\alias{sample_names}
\title{Samples random names}
\usage{
sample_names(n = 1, type = "neutral", seed = NULL)
}
\arguments{
\item{n}{Number of names}

\item{type}{"neutral", "male", "female", or "mixed"}

\item{seed}{A seed for the random number generator.}
}
\value{
A character vector with random names
}
\description{
Samples random names
}
\examples{
sample_names(3)
}
