% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{package}
\name{scan-package}
\alias{scan-package}
\alias{scan}
\title{Single-Case Data Analyses}
\description{
A collection of procedures for analysing, visualising, and managing single-case data.
}
\author{
Juergen Wilbert [aut, cre]
}
\keyword{package}
