\name{scan-package}
\alias{scan-package}
\alias{scan}
\docType{package}
\title{Single-Case Data Analyses}
\description{
A package for analyzing single-case data of from AB-designs. Most procedures also support multiple-baseline AB-designs.}
\details{
\tabular{ll}{
Package: \tab scan\cr
Type: \tab Package\cr
Version: \tab 0.17\cr
Date: \tab 2016-07-04\cr
License: \tab GPL (version 2 or newer)\cr
}
}

\author{Juergen Wilbert [aut, cre]}