\name{scan-package}
\alias{scan-package}
\alias{scan}
\alias{print.sc}
\docType{package}
\title{Single-Case Data Analyses}
\description{
A package to analyze single-case data.

Most procedures also support multiple-baseline designs.}
\details{
\tabular{ll}{
Package: \tab scan\cr
Type: \tab Package\cr
Version: \tab 0.26\cr
Date: \tab 2017-06-19\cr
License: \tab GPL (version 2 or newer)\cr
}
}

\author{Juergen Wilbert [aut, cre]}

\keyword{package}