\name{scan-package}
\alias{scan-package}
\alias{scan}
\alias{print.sc}
\docType{package}
\title{Single-Case Data Analyses
}
\description{
A collection of procedures for analysing single-case data of an AB-design. Some procedures support multiple-baseline designs.}
\details{
\tabular{ll}{
Package: \tab scan\cr
Type: \tab Package\cr
Version: \tab 0.17\cr
Date: \tab 2016-07-04\cr
License: \tab GPL (version 2 or newer)\cr
}
}
\author{
Juergen Wilbert

Department of Teacher Education

Faculty of Humanities

University of Potsdam

 http://www.uni-potsdam.de/inklusion/forschungsmethodenunddiagnostik/wilbert/projekte/single-case-data-analysis-with-r.html

Maintainer: Juergen Wilbert <jwilbert@uni-potsdam.de>
}
