\name{scan-package}
\alias{scan-package}
\alias{scan}

\alias{print.sc}
\alias{estimateSC}
\alias{keepphasesSC}
\alias{newrankSC}
\alias{recodeSCDF}
\alias{scaleSC}


\docType{package}
\title{Single-Case Data Analyses}
\description{
A package to analyze single-case data.

Most procedures also support multiple-baseline designs.}
\details{
\tabular{ll}{
Package: \tab scan\cr
Type: \tab Package\cr
Version: \tab 0.33\cr
Date: \tab 2017-08-08\cr
License: \tab GPL (version 2 or newer)\cr
}
}

\author{Juergen Wilbert [aut, cre]}

\keyword{package}