\name{scan-package}
\alias{scan-package}
\alias{scan}
\alias{print.sc}
\docType{package}
\title{Single-Case Data Analyses}
\description{
A package to analyze single-case data of from AB-designs.

Most procedures also support multiple-baseline AB-designs.}
\details{
\tabular{ll}{
Package: \tab scan\cr
Type: \tab Package\cr
Version: \tab 0.22\cr
Date: \tab 2016-11-09\cr
License: \tab GPL (version 2 or newer)\cr
}
}

\author{Juergen Wilbert [aut, cre]}

\keyword{package}