% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf_attr.R
\name{scdf_attr}
\alias{scdf_attr}
\alias{scdf_attr<-}
\title{Set and get scdf attributes}
\usage{
scdf_attr(x, var)

scdf_attr(x, var) <- value
}
\arguments{
\item{x}{Variable}

\item{var}{Attribute}

\item{value}{set value}
}
\value{
Attribute value
}
\description{
Set and get scdf attributes
}
