% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot.R
\name{scplot}
\alias{scplot}
\alias{set_style}
\alias{add_statline}
\alias{set_xlabel}
\alias{set_ylabel}
\alias{set_xaxis}
\alias{set_yaxis}
\alias{add_title}
\alias{add_marks}
\alias{set_phasenames}
\alias{set_casenames}
\alias{add_text}
\alias{add_arrow}
\alias{add_grid}
\alias{set_background}
\alias{add_frame}
\alias{add_box}
\alias{add_labels}
\alias{set_line}
\alias{set_dots}
\alias{add_ridge}
\alias{set_seperator}
\title{Plot single-case data}
\usage{
scplot(scdf)

set_style(object, style, ...)

add_statline(object, stat, ...)

set_xlabel(object, label, col, size)

set_ylabel(object, label, col, size, orientation)

set_xaxis(object, limits, increment, col, size)

set_yaxis(object, limits, col, size)

add_title(object, label, col, size, font)

add_marks(object, case, positions, col = "red", size = 1, shape = 1)

set_phasenames(object, ..., col, size)

set_casenames(object, ..., col, side, size)

add_text(object, case = 1, x, y, label, col = NULL, size = NULL, angle = 0)

add_arrow(object, case = 1, x0, y0, x1, y1, length = 0.1, ...)

add_grid(object, type, width, col)

set_background(object, col)

add_frame(object, col = "black")

add_box(object, col = "black", type = "solid", width = 1)

add_labels(
  object,
  col = "black",
  size = 0.6,
  offset = 0.4,
  position = 3,
  round = 1
)

set_line(object, col, width, type)

set_dots(object, col, size, shape)

add_ridge(object, col = "grey98")

set_seperator(object, col, width, type, extent)
}
\arguments{
\item{scdf}{A single-case data frame.}

\item{...}{Further arguments passed to the function.}

\item{label}{Character string.}

\item{x}{x position}

\item{y}{y position}

\item{lines}{A list defining one or multiple lines or curves to be
plotted. The argument is passed as a list (e.g., \code{list(type = "median")}).
Some of the procedures can be refined with an additional argument (e.g.,
\code{lines = list(type = "mean", trim = 0.2)} adds a 20\% trimmed mean line.
For multiple lines, provide a list element for each line (e.g., \code{list( list(type = "median", col = "red"), list(type = "trend", col = "blue"))}. 
Possible lines are: \itemize{
\item\code{"median"} Separate lines for phase A and B medians.
\item\code{"mean"} Separate lines for phase A and B means. By default it is
10\%-trimmed. Other trims can be set, using a trim parameter (e.g.,
\code{lines = list(type = "mean", trim = 0.2)} draws a 20\%-trimmed mean line).
\item\code{"trend"} Separate lines for phase A and B trends.
\item\code{"trendA"} OLS trend line for phase A, extrapolated throughout phase B.
\item\code{"trendA_bisplit"} Split middle (bi-split) trend line for phase A, extrapolated throughout phase B.
\item\code{"trendA_trisplit"} Tukey tri-split trend line for phase A, extrapolated throughout phase B.
\item\code{"maxA/minA"} Line at the level of the highest or lowest phase A score.
\item\code{"medianA"} Line at the phase A median score.  \item\code{"meanA"}
Line at the phase A 10\%-trimmed mean score. Apply a different trim, by
using the additional argument (e.g., \code{lines = list(type = "meanA", trim = 0.2)}).
\item\code{"plm"} Regression lines for piecewise linear regression model.
\item\code{"plm.ar"} Regression lines for piecewise autoregression model.
The lag is specified like this: \code{lines = list(type = "plm.ar", ar = 2)}. Default lag is set to 2.
\item\code{"movingMean"} Draws a moving mean curve, with a specified lag:
\code{lines = list(type = "movingMean", lag = 2)}. Default is a lag 1 curve.
\item\code{"movingMedian"} Draws a moving median curve, with a specified
lag: \code{lines = list(type = "movingMedian", lag = 3)}. Default is a lag 1 curve.
\item\code{"loreg"} Draws a non-parametric local regression line. The
proportion of data influencing each data point can be specified using
\code{lines = list(type = "loreg"m f = 0.66)}. The default is 0.5.  \item\code{"lty"}
Use this argument to define the line type. Examples are: \code{"solid"},
\code{"dashed"}, \code{"dotted"}.  \item\code{"lwd"} Use this argument to
define the line's thickness, e.g., \code{lwd = 4}.  \item\code{"col"} Use
this argument to define the line's color, e.g., \code{col = "red"}.  }}

\item{marks}{A list of parameters defining markings of certain data points.
\itemize{ \item\code{"positions"} A vector or a list of vectors indicating
measurement-times to be highlighted. In case of a vector, the marked
measurement-times are the same for all plotted cases. In case of a list of
vectors, marks are set differently for each case. The list must have the
same length as there are cases in the data file.  \item\code{"col"} Color of
the marks.  \item\code{"cex"} Size of the marks.  } Use for example
\code{marks = list(positions = c(1, 8, 15), col = "red", cex = 3)} to make
the MTs one, eight and 18 appear big and red.}
}
\value{
An scdf plot object that creates a plot when printed.
}
\description{
This function provides a plot of a single-case or multiple
single-cases.
}
