% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot.R
\name{scplot}
\alias{scplot}
\alias{add_theme}
\alias{set_theme_element}
\alias{add_statline}
\alias{set_xlabel}
\alias{set_ylabel}
\alias{set_xaxis}
\alias{set_yaxis}
\alias{add_title}
\alias{add_caption}
\alias{add_marks}
\alias{set_phasenames}
\alias{set_casenames}
\alias{add_text}
\alias{add_arrow}
\alias{add_grid}
\alias{set_background}
\alias{add_frame}
\alias{add_box}
\alias{add_labels}
\alias{set_line}
\alias{set_dots}
\alias{add_ridge}
\alias{set_seperator}
\alias{add_dataline}
\alias{add_legend}
\alias{add_margins}
\title{Plot single-case data}
\usage{
scplot(data)

add_theme(object, theme, ...)

set_theme_element(object, ...)

add_statline(
  object,
  stat,
  colour = NULL,
  width = NULL,
  type = NULL,
  variable = ".dvar",
  ...
)

set_xlabel(object, label, colour, size)

set_ylabel(object, label, colour, size, line, orientation)

set_xaxis(
  object,
  limits,
  increment,
  increment_from,
  colour,
  size,
  line,
  positions
)

set_yaxis(object, limits, colour, size, increment, increment_from, positions)

add_title(object, label, colour, size, font, align, parse)

add_caption(object, label, colour, size, font, align, wrap, margin, parse)

add_marks(
  object,
  case = 1,
  positions,
  colour = "red",
  size = 1,
  shape = 1,
  variable = ".dvar"
)

set_phasenames(object, ..., colour, size)

set_casenames(object, ..., colour, side, size)

add_text(object, case = 1, x, y, label, colour = NULL, size = NULL, angle = 0)

add_arrow(object, case = 1, x0, y0, x1, y1, length = 0.1, colour = NULL, ...)

add_grid(object, type, width, colour)

set_background(object, colour)

add_frame(object, colour = "black", size = 1, type = "solid")

add_box(object, colour = "black", type = "solid", width = 1)

add_labels(
  object,
  colour = "black",
  size = 0.6,
  offset = 0.4,
  position = 3,
  round = 1
)

set_line(object, colour, width, type, variable)

set_dots(object, colour, size, shape, variable)

add_ridge(object, colour = "grey98")

set_seperator(object, colour, width, type, extent, label, size)

add_dataline(
  object,
  variable,
  colour = NULL,
  width = NULL,
  type = NULL,
  colour_dots = NULL,
  shape = NULL,
  size = NULL
)

add_legend(object, labels = ".default", x, y)

add_margins(object, outer, inner)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{theme}{A character string with a predefined graphical theme. 
Possible values: default, yaxis, tiny, small, big, chart, ridge, 
annotate, grid, grid2, dark, nodots, sienna, phase_color, phase_shade}

\item{...}{various style parameter}

\item{stat}{A character string defining a line or curve to be
plotted. Possible values: "median", "mean", "trend", "trendA", 
"trendA_bisplit", "trendA_trisplit", "maxA", "minA", "meanA", "medianA", 
"plm", "movingMean", "movingMedian", "loreg"}

\item{colour}{A character string or a number defining the color of an element.}

\item{width}{A number defining the line width}

\item{type}{A character string with the line type: "solid", "dashed", "dotted"}

\item{label}{Character string.}

\item{size}{A number deifning the size of an element.}

\item{orientation}{of the label: 0 = vertical; 1 = horizontal}

\item{limits}{Lower and upper limits of the axis (e.g., \code{limits = c(0,
20)} sets the axis to a scale from 0 to 20). With multiple single-cases
you can use \code{limits = c(0, NA)} to scale the axis from 0 to the maximum
of each case. \code{limits} is not set by default, which makes \code{scan} set
a proper scale based on the given data.}

\item{increment}{An integer. Increment of the x-axis. 1 :each mt value will be printed, 2 : every other value, 3 : every third values etc.}

\item{positions}{Either a vector indicating the dot to be highlighted or a 
character string with a logical expression (e.g. values < mean(values))}

\item{align}{Character string. One of "left", "right", "center"}

\item{parse}{If TRUE, the label is interpreted as an expression. Default = FALSE.}

\item{wrap}{Number that defines the maximum characters per line before a break. 
If set to FALSE, no automatic linebreak is set.}

\item{case}{Numerical vector with the csae number or character string "all" for all cases.}

\item{shape}{Number. See pch graphical parameter on par help page.}

\item{x}{x position}

\item{y}{y position}

\item{x0}{Origin x position of the line.}

\item{y0}{Origin y position of the line.}

\item{x1}{End x position of the line.}

\item{y1}{End y position of the line.}

\item{length}{Size of the aroow angels.}

\item{offset}{Number with shift from exact position of the label.}

\item{position}{Number 1, 2, 3,or 4, indicate positions below, to the left of, above and to the right of the exact value coordinate.}

\item{round}{Number of digits of the labels.}

\item{extent}{A number between 0 and 1 given the proportion of the plot that is covert by the line or character string "full" or "scale".}

\item{outer}{Vector with four values for the extension of the inner margins
(negative numbers for smaller margins): c(bottom, left, top, right)}
}
\value{
An scdf plot object that creates a plot when printed.
}
\description{
This function provides a plot of a single-case or multiple
single-cases.
}
