% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cases.R
\name{select_cases}
\alias{select_cases}
\title{Select a subset of cases}
\usage{
select_cases(scdf, ...)
}
\arguments{
\item{scdf}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{...}{Selection criteria. Either numeric or as characters.}
}
\value{
An scdf with a subset of cases
}
\description{
Select a subset of cases
}
\examples{
select_cases(exampleAB, "Johanna", "Karolina")
select_cases(exampleAB, 1,2)
select_cases(exampleAB, "-Johanna")
}
