% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_phases.R
\name{select_phases}
\alias{select_phases}
\title{Select and combine phases for overlap analyses}
\usage{
select_phases(data, A, B)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link[=scdf]{scdf()}} to learn about this
format.}

\item{A}{Selection of the A phase}

\item{B}{Selection of the B phase}
}
\value{
An scdf with selected phases
}
\description{
Useful when working with \%>\% operators.
}
\examples{
exampleA1B1A2B2_zvt \%>\% 
  select_phases(A = c(1, 3), B = c(2, 4)) \%>\%
  overlap()
}
