% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_phases.R
\name{select_phases}
\alias{select_phases}
\title{Select phases for overlap analyses}
\usage{
select_phases(data, A, B, ...)
}
\arguments{
\item{data}{}

\item{A}{Selection of the A phase}

\item{B}{Selection of the B phase}
}
\value{
An scdf with selected phases
}
\description{
Useful when working with %>% operators.
}
