% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_var.R
\name{set_vars}
\alias{set_vars}
\title{Set analysis variables in an scdf}
\usage{
set_vars(data, dv, mt, phase)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{dv}{Character string. Name of the dependent variable.}

\item{mt}{Character string. Name of the measurement-time variable.}

\item{phase}{Character string. Name of the phase variable.}
}
\description{
Set analysis variables in an scdf
}
