% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyscan.R
\name{shinyscan}
\alias{shinyscan}
\title{A Shiny app for scan}
\usage{
shinyscan(
  scdf = NULL,
  quiet = TRUE,
  browser = c("external", "viewer"),
  theme = "cerulean",
  ...
)
}
\arguments{
\item{scdf}{If you provide an \emph{scdf} here, it will be loaded at startup.}

\item{quiet}{If TRUE (default) does not report shiny messages in the console.}

\item{browser}{c("external","viewer")}

\item{theme}{Bootstrap 5 theme. Default is "cerulean".}

\item{\dots}{Further arguments passed to the \code{shiny::runApp()} function.}
}
\description{
Run a Shiny app with most of the scan functions.
}
\details{
This function launches a shiny application.
You need to have \code{scplot} and \code{shiny} installed.
These packages are suggested but not necessarily installed along with scan.
\code{shinyscan()} will ask to install missing packages.
}
