% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R, R/print-export-smd.R
\name{smd}
\alias{smd}
\alias{print.sc_smd}
\alias{export.sc_smd}
\title{Standardized mean differences for single-case data}
\usage{
smd(data, dvar, pvar, phases = c(1, 2))

\method{print}{sc_smd}(x, digits = "auto", ...)

\method{export}{sc_smd}(
  object,
  caption = NA,
  footnote = NA,
  filename = NA,
  select = c("Case", `Mean A` = "mA", `Mean B` = "mB", `SD A` = "sdA", `SD B` = "sdB",
    `SD Cohen` = "sd cohen", `SD Hedges` = "sd hedges", "Glass' delta", "Hedges' g",
    "Hedges' g correction", "Hedges' g durlak correction", "Cohen's d"),
  round = 2,
  decimals = 2,
  flip = FALSE,
  ...
)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link[=scdf]{scdf()}} to learn about this
format.}

\item{dvar}{Character string with the name of the dependent variable.
Defaults to the attributes in the scdf file.}

\item{pvar}{Character string with the name of the phase variable. Defaults to
the attributes in the scdf file.}

\item{phases}{A vector of two characters or numbers indicating the two phases
that should be compared. E.g., \code{phases = c("A","C")} or \code{phases = c(2,4)}
for comparing the second to the fourth phase. Phases could be combined by
providing a list with two elements. E.g., \code{phases = list(A = c(1,3), B = c(2,4))} will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default
is \code{phases = c(1,2)}.}

\item{x}{Object}

\item{digits}{The minimum number of significant digits to be use.
If set to "auto" (default), values are predefined.}

\item{...}{Further arguments passed to the function.}

\item{object}{An scdf or an object exported from a scan function.}

\item{caption}{Character string with table caption. If left NA (default) a
caption will be created based on the exported object.}

\item{footnote}{Character string with table footnote. Several strings are
combined into a footnote of several lines. If left NA (default) a footnote
will be created based on the exported object. \code{NULL} or \code{""} suppress the
footnote.}

\item{filename}{String containing the file name. If a filename is given the
output will be written to that file.}

\item{select}{A character vector containing the names of the variables to be
included. If the vector is named, the variables will be renamed
accordingly.}

\item{round}{Integer passed to the digits argument used to round values.}

\item{decimals}{Decimal places that are reported.}

\item{flip}{If TRUE, some objects are exported with rows and columns flipped.}
}
\description{
The \code{smd()} function provides various standardized mean effect sizes for
single-case data.
}
\details{
It computes 'Cohen's d', 'Hedges' g', 'Hedges' g correction',
'Hedges' g durlak correction', and 'Glass' delta' for each single-case
included in an scdf.

'sd cohen' is the (unweigted) average of the variance of phase A and
B. 'sd Hedges' is the weighted average of the variance of phase A and B
(with a degrees of freedom correction). 'Hedges' g' is the mean difference
divided by 'sd Hedges'. 'Hedges' g correction' and 'Hedges' g durlak
correction' are two approaches of correcting Hedges' g for small sample
sizes. 'Glass' delta' is the mean difference divided by the standard
deviation of the A-phase. 'Cohens d' is the mean difference divided by 'sd
cohen'.
}
\section{Functions}{
\itemize{
\item \code{print(sc_smd)}: Print results

\item \code{export(sc_smd)}: Export results as html table (see \code{\link[=export]{export()}})

}}
\examples{
smd(exampleAB)
}
\seealso{
\code{\link[=overlap]{overlap()}}, \code{\link[=describe]{describe()}}
}
\author{
Juergen Wilbert
}
