% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_cases}
\alias{smooth_cases}
\alias{smoothSC}
\title{Smoothing single-case data}
\usage{
smooth_cases(data, dvar, mvar, FUN = "movingMedian", intensity = NULL)

smoothSC(...)
}
\arguments{
\item{data}{A single-case data frame. See \code{\link{scdf}} to learn about this format.}

\item{dvar}{Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.}

\item{mvar}{Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.}

\item{FUN}{Function determining the smoothed scores. Default \code{FUN =
"movingMedian"} is a moving Median function. Further possible values are:
\code{"movingMean"} and a non-parametric \code{"localRegression"}.}

\item{intensity}{For \code{FUN = "movingMedian"} and \code{"movingMean"} it
is the lag used for computing the average. Default is \code{intensity = 1}.
In case of \code{FUN = "localRegression"} it is the proportion of
surrounding data influencing each data point, which is \code{intensity =
0.2} by default.}

\item{...}{Further arguments passed to the function.}
}
\value{
Returns a data frame (for each single-case) with smoothed data
points. See \code{\link{scdf}} to learn about the format of these data
frames.
}
\description{
The \code{smooth_cases} function provides procedures to smooth single-case data
(i.e., to eliminate noise). A moving average function (mean- or
median-based) replaces each data point by the average of the surrounding
data points step-by-step. With a local regression function, each data point
is regressed by its surrounding data points.
}
\examples{

## Use the three different smoothing functions and compare the results
study <- c(
  "Original" = Huber2014$Berta,
  "Moving Median" = smooth_cases(Huber2014$Berta, FUN = "movingMedian"),
  "Moving Mean" = smooth_cases(Huber2014$Berta, FUN = "movingMean"),
  "Local Regression" = smooth_cases(Huber2014$Berta, FUN = "localRegression")
)
plot(study)
}
\seealso{
Other data manipulation functions: 
\code{\link{as.data.frame.scdf}()},
\code{\link{fill_missing}()},
\code{\link{outlier}()},
\code{\link{ranks}()},
\code{\link{shift}()},
\code{\link{standardize}()},
\code{\link{truncate_phase}()}
}
\author{
Juergen Wilbert
}
\concept{data manipulation functions}
\keyword{manip}
