% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.scdf.R
\name{subset.scdf}
\alias{subset.scdf}
\title{Subset cases, rows, and variables}
\usage{
\method{subset}{scdf}(x, filter, select, cases, ...)
}
\arguments{
\item{x}{An scdf object.}

\item{filter}{Logical expression indicating rows to keep: missing values are taken as false.}

\item{select}{Expression, indicating columns to select from an scdf.}

\item{cases}{Expression, indicating cases to keep from an scdf.}

\item{...}{[use not implemented yet]}
}
\value{
An scdf with cases filtered by the filter expression.
}
\description{
Subset cases, rows, and variables
}
\examples{
subset(exampleAB, (values < 60 & phase == "A") | (values >= 60 & phase == "B"))
subset(exampleAB_add, select = c(-cigarrets, -depression))
subset(exampleAB, cases = c(Karolina, Johanna))
subset(exampleA1B1A2B2, phase \%in\% c("A1", "B2"), cases = Pawel:Moritz)
}
