% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdf-functions.R
\name{$.scdf}
\alias{$.scdf}
\alias{[.scdf}
\title{Select a scdf}
\usage{
\method{$}{scdf}(x, i)

\method{[}{scdf}(x, i)
}
\arguments{
\item{x}{A scdf object}

\item{i}{A case name from x}
}
\value{
A scdf
}
\description{
Select a scdf
}
