% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scdf.R
\name{summary.scdf}
\alias{summary.scdf}
\title{Summary function for an scdf object}
\usage{
\method{summary}{scdf}(object, all_cases = FALSE, ...)
}
\arguments{
\item{object}{scdf}

\item{all_cases}{IF TRUE, more that 10 cases are summarized}

\item{...}{not in use}
}
\description{
Provides a summary of an \code{scdf} object, including the number of cases,
measurements per case, and design information.
}
\details{
The summary includes:
\itemize{
\item Total number of cases in the \code{scdf}.
\item A table listing each case with the number of measurements and design.
\item Variable names with annotations for phase, measurement-time, and dependent variable.
\item Additional information and author details if available.
}
}
\author{
Juergen Wilbert
}
