% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scdf.R, R/print-export-scdf-summary.R
\name{summary.scdf}
\alias{summary.scdf}
\alias{print.scdf_summary}
\alias{export.scdf_summary}
\title{Summary function for an scdf object}
\usage{
\method{summary}{scdf}(object, all_cases = FALSE, ...)

\method{print}{scdf_summary}(x, all_cases = NULL, ...)

\method{export}{scdf_summary}(object, caption = NA, footnote = NA, filename = NA, round = 2, ...)
}
\arguments{
\item{object}{An scdf object}

\item{all_cases}{If TRUE, more than 10 cases are summarized}

\item{...}{not in use}

\item{x}{An object of class \code{scdf_summary}}

\item{caption}{Character string with table caption. If left NA (default) a
caption will be created based on the exported object.}

\item{footnote}{Character string with table footnote. Several strings are
combined into a footnote of several lines. If left NA (default) a footnote
will be created based on the exported object. \code{NULL} or \code{""} suppress the
footnote.}

\item{filename}{String containing the file name. If a filename is given the
output will be written to that file.}

\item{round}{Integer passed to the digits argument used to round values.}
}
\value{
An object of class \code{scdf_summary}.
}
\description{
Provides a summary of an \code{scdf} object, including the number of cases,
measurements per case, and design information.
}
\details{
The summary includes:
\itemize{
\item Total number of cases in the \code{scdf}.
\item A table listing each case with the number of measurements and design.
\item Variable names with annotations for phase, measurement-time, and dependent variable.
\item Additional information and author details if available.
}

\code{summary()} returns the summary object; the output is written by its print
method. So \code{summary(scdf)} shows the summary at the console as before, while
\code{export(summary(scdf))} and an assignment stay silent.
}
\section{Functions}{
\itemize{
\item \code{print(scdf_summary)}: Print the summary

\item \code{export(scdf_summary)}: Export the summary as html table (see \code{\link[=export]{export()}})

}}
\author{
Juergen Wilbert
}
