% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scdf.R
\name{summary.scdf}
\alias{summary.scdf}
\title{Summary function for an scdf}
\usage{
\method{summary}{scdf}(object, all_cases = FALSE, ...)
}
\arguments{
\item{object}{scdf}

\item{all_cases}{IF TRUE, more that 10 cases are summarized}

\item{...}{not in use}
}
\description{
Summary function for an scdf
}
