% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scdf.R
\name{summary.scdf}
\alias{summary.scdf}
\title{Summary functio for a scdf}
\usage{
\method{summary}{scdf}(object, ...)
}
\arguments{
\item{object}{scdf}

\item{...}{not in use}
}
\description{
Summary functio for a scdf
}
