% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateSC.R
\name{truncateSC}
\alias{truncateSC}
\title{Truncate single-case data}
\usage{
truncateSC(data, truncate = list(A = c(0, 0), B = c(0, 0)))
}
\arguments{
\item{data}{A single-case data frame or a list of single-case data frames.
See \code{\link{scdf}} to learn about this format.}

\item{truncate}{A list with a vector of two (beginning and end) values for each phase defining the number of data points to be deleted.
For lists of single-case data frames, the truncation is adapted to the length 
of each phase for each single case.}
}
\value{
A truncated data frame (for each single-case).
}
\description{
This function truncates data points at the beginning and / or end each phase.
}
\examples{

# Truncate the first two data points of both phases and compare the two data sets
study <- c(byHeart2011[1], truncateSC(byHeart2011[1], list(A = c(2,0), B = c(2,0))))
names(study) <- c("Original","Selected")
plot(study)

}
\seealso{
Other data manipulation functions: \code{\link{fillmissingSC}},
  \code{\link{longSCDF}}, \code{\link{rankSC}},
  \code{\link{scaleSC}}, \code{\link{smoothSC}}
}
\author{
Juergen Wilbert
}
\concept{data manipulation functions}
\keyword{manip}
