\name{writeSC}
\alias{writeSC}
\title{
Export data into a csv file.
}
\description{
Data are restructured and exported into a csv file for later use with other software.
}
\usage{
writeSC(dat, filename, sep = ",", dec = ".", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{
Data file to be exported.
}
  \item{filename}{
Target file name.
}
  \item{sep}{
Variable separator.
}
  \item{dec}{
Decimal sign.
}
  \item{\dots}{
Further arguments passed to write.table.
}
}
\author{Juergen Wilbert}
\examples{
#
}
